GitHub

Description

This module re-exports all request constructors and data definitions from this package.

See GitHub.Request module for executing Request, in short use github request, for example

github userInfoForR
  :: AuthMethod am => am -> Name User -> IO (Either Error User)

The missing endpoints lists show which endpoints we know are missing, there might be more.

Activity

See https://developer.github.com/v3/activity/

Events

See https://developer.github.com/v3/activity/events/

repositoryEventsR :: Name Owner -> Name Repo -> FetchCount -> Request 'RO (Vector Event)

List repository events. See https://developer.github.com/v3/activity/events/#list-repository-events

userEventsR :: Name User -> FetchCount -> Request 'RO (Vector Event)

List user public events. See https://developer.github.com/v3/activity/events/#list-public-events-performed-by-a-user

Notifications

See https://developer.github.com/v3/activity/notifications/

getNotificationsR :: FetchCount -> Request 'RA (Vector Notification)

List your notifications. See https://developer.github.com/v3/activity/notifications/#list-your-notifications

markNotificationAsReadR :: Id Notification -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Mark a thread as read. See https://developer.github.com/v3/activity/notifications/#mark-a-thread-as-read

markAllNotificationsAsReadR :: GenRequest ('MtUnit :: MediaType Type) 'RW ()

Mark as read. See https://developer.github.com/v3/activity/notifications/#mark-as-read

Starring

See https://developer.github.com/v3/activity/starring/

Missing endpoints:

Check if you are starring a repository

stargazersForR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector SimpleUser)

List Stargazers. See https://developer.github.com/v3/activity/starring/#list-stargazers

reposStarredByR :: forall (k :: RW). Name Owner -> FetchCount -> Request k (Vector Repo)

List repositories being starred. See https://developer.github.com/v3/activity/starring/#list-repositories-being-starred

myStarredR :: FetchCount -> Request 'RA (Vector Repo)

All the repos starred by the authenticated user. See https://developer.github.com/v3/activity/starring/#list-repositories-being-starred

myStarredAcceptStarR :: FetchCount -> GenRequest ('MtStar :: MediaType Type) 'RA (Vector RepoStarred)

All the repos starred by the authenticated user. See https://developer.github.com/v3/activity/starring/#alternative-response-with-star-creation-timestamps-1

starRepoR :: Name Owner -> Name Repo -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Star a repo by the authenticated user. See https://developer.github.com/v3/activity/starring/#star-a-repository

unstarRepoR :: Name Owner -> Name Repo -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Unstar a repo by the authenticated user. See https://developer.github.com/v3/activity/starring/#unstar-a-repository

Watching

See https://developer.github.com/v3/activity/

Missing endpoints:

Query a Repository Subscription

Set a Repository Subscription

watchersForR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector SimpleUser)

List watchers. See https://developer.github.com/v3/activity/watching/#list-watchers

reposWatchedByR :: forall (k :: RW). Name Owner -> FetchCount -> Request k (Vector Repo)

List repositories being watched. See https://developer.github.com/v3/activity/watching/#list-repositories-being-watched

unwatchRepoR :: Name Owner -> Name Repo -> Request 'RW ()

Stop watching repository. See https://docs.github.com/en/rest/reference/activity#delete-a-repository-subscription

Gists

See https://developer.github.com/v3/gists/

Missing endpoints:

Query a specific revision of a gist

Edit a gist

List gist commits

Check if a gist is starred

Fork a gist

List gist forks

gistsR :: forall (k :: RW). Name Owner -> FetchCount -> Request k (Vector Gist)

List gists. See https://developer.github.com/v3/gists/#list-gists

gistR :: forall (k :: RW). Name Gist -> Request k Gist

Query a single gist. See https://developer.github.com/v3/gists/#get-a-single-gist

createGistR :: NewGist -> Request 'RW Gist

Create a new gist See https://docs.github.com/rest/reference/gists#create-a-gist

starGistR :: Name Gist -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Star a gist by the authenticated user. See https://developer.github.com/v3/gists/#star-a-gist

unstarGistR :: Name Gist -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Unstar a gist by the authenticated user. See https://developer.github.com/v3/gists/#unstar-a-gist

deleteGistR :: Name Gist -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a gist by the authenticated user. See https://developer.github.com/v3/gists/#delete-a-gist

Comments

See https://developer.github.com/v3/gists/comments/

Missing endpoints: * Create a comment * Edit a comment * Delete a comment

commentsOnR :: forall (k :: RW). Name Gist -> FetchCount -> Request k (Vector GistComment)

List comments on a gist. See https://developer.github.com/v3/gists/comments/#list-comments-on-a-gist

gistCommentR :: forall (k :: RW). Id GistComment -> Request k GistComment

Query a single comment. See https://developer.github.com/v3/gists/comments/#get-a-single-comment

Git Data

See https://developer.github.com/v3/git/

Blobs

See https://developer.github.com/v3/git/blobs/

blobR :: forall (k :: RW). Name Owner -> Name Repo -> Name Blob -> Request k Blob

Query a blob. See https://developer.github.com/v3/git/blobs/#get-a-blob

Commits

See https://developer.github.com/v3/git/commits/

gitCommitR :: forall (k :: RW). Name Owner -> Name Repo -> Name GitCommit -> Request k GitCommit

Query a commit. See https://developer.github.com/v3/git/commits/#get-a-commit

References

See https://developer.github.com/v3/git/refs/

referenceR :: forall (k :: RW). Name Owner -> Name Repo -> Name GitReference -> Request k GitReference

A single reference -- | Query a reference. See https://developer.github.com/v3/git/refs/#get-a-reference

referencesR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector GitReference)

Query all References. See https://developer.github.com/v3/git/refs/#get-all-references

createReferenceR :: Name Owner -> Name Repo -> NewGitReference -> Request 'RW GitReference

Create a reference. See https://developer.github.com/v3/git/refs/#create-a-reference

deleteReferenceR :: Name Owner -> Name Repo -> Name GitReference -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a reference. See https://developer.github.com/v3/git/refs/#delete-a-reference

namespacedReferencesR :: forall (k :: RW). Name Owner -> Name Repo -> Text -> Request k [GitReference]

Query namespaced references. See https://developer.github.com/v3/git/refs/#get-all-references

Trees

See https://developer.github.com/v3/git/trees/

treeR :: forall (k :: RW). Name Owner -> Name Repo -> Name Tree -> Request k Tree

Query a Tree. See https://developer.github.com/v3/git/trees/#get-a-tree

nestedTreeR :: forall (k :: RW). Name Owner -> Name Repo -> Name Tree -> Request k Tree

Query a Tree Recursively. See https://developer.github.com/v3/git/trees/#get-a-tree-recursively

Issues

See https://developer.github.com/v3/issues/

currentUserIssuesR :: IssueMod -> FetchCount -> Request 'RA (Vector Issue)

See https://developer.github.com/v3/issues/#list-issues.

organizationIssuesR :: forall (k :: RW). Name Organization -> IssueMod -> FetchCount -> Request k (Vector Issue)

See https://developer.github.com/v3/issues/#list-issues.

issueR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> Request k Issue

Query a single issue. See https://developer.github.com/v3/issues/#get-a-single-issue

issuesForRepoR :: forall (k :: RW). Name Owner -> Name Repo -> IssueRepoMod -> FetchCount -> Request k (Vector Issue)

List issues for a repository. See https://developer.github.com/v3/issues/#list-issues-for-a-repository

createIssueR :: Name Owner -> Name Repo -> NewIssue -> Request 'RW Issue

Create an issue. See https://developer.github.com/v3/issues/#create-an-issue

editIssueR :: Name Owner -> Name Repo -> IssueNumber -> EditIssue -> Request 'RW Issue

Edit an issue. See https://developer.github.com/v3/issues/#edit-an-issue

Comments

See https://developer.github.com/v3/issues/comments/

commentR :: forall (k :: RW). Name Owner -> Name Repo -> Id Comment -> Request k IssueComment

Query a single comment. See https://developer.github.com/v3/issues/comments/#get-a-single-comment

commentsR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> FetchCount -> Request k (Vector IssueComment)

List comments on an issue. See https://developer.github.com/v3/issues/comments/#list-comments-on-an-issue

createCommentR :: Name Owner -> Name Repo -> IssueNumber -> Text -> Request 'RW Comment

Create a comment. See https://developer.github.com/v3/issues/comments/#create-a-comment

deleteCommentR :: Name Owner -> Name Repo -> Id Comment -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a comment. See https://developer.github.com/v3/issues/comments/#delete-a-comment

editCommentR :: Name Owner -> Name Repo -> Id Comment -> Text -> Request 'RW Comment

Edit a comment. See https://developer.github.com/v3/issues/comments/#edit-a-comment

Events

See https://developer.github.com/v3/issues/events/

eventsForIssueR :: forall (k :: RW). Name Owner -> Name Repo -> Id Issue -> FetchCount -> Request k (Vector IssueEvent)

List events for an issue. See https://developer.github.com/v3/issues/events/#list-events-for-an-issue

eventsForRepoR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector IssueEvent)

List events for a repository. See https://developer.github.com/v3/issues/events/#list-events-for-a-repository

eventR :: forall (k :: RW). Name Owner -> Name Repo -> Id IssueEvent -> Request k IssueEvent

Query a single event. See https://developer.github.com/v3/issues/events/#get-a-single-event

Labels

See https://developer.github.com/v3/issues/labels/

labelsOnRepoR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector IssueLabel)

List all labels for this repository. See https://developer.github.com/v3/issues/labels/#list-all-labels-for-this-repository

labelR :: forall (k :: RW). Name Owner -> Name Repo -> Name IssueLabel -> Request k IssueLabel

Query a single label. See https://developer.github.com/v3/issues/labels/#get-a-single-label

createLabelR :: Name Owner -> Name Repo -> NewIssueLabel -> Request 'RW IssueLabel

Create a label. See https://developer.github.com/v3/issues/labels/#create-a-label

updateLabelR

Arguments

:: Name Owner

-> Name Repo

-> Name IssueLabel
old label name

-> UpdateIssueLabel
new label

-> Request 'RW IssueLabel

Update a label. See https://developer.github.com/v3/issues/labels/#update-a-label

deleteLabelR :: Name Owner -> Name Repo -> Name IssueLabel -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a label. See https://developer.github.com/v3/issues/labels/#delete-a-label

labelsOnIssueR :: forall (k :: RW). Name Owner -> Name Repo -> Id Issue -> FetchCount -> Request k (Vector IssueLabel)

List labels on an issue. See https://developer.github.com/v3/issues/labels/#list-labels-on-an-issue

addLabelsToIssueR :: Foldable f => Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> Request 'RW (Vector IssueLabel)

Add lables to an issue. See https://developer.github.com/v3/issues/labels/#add-labels-to-an-issue

removeLabelFromIssueR :: Name Owner -> Name Repo -> Id Issue -> Name IssueLabel -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Remove a label from an issue. See https://developer.github.com/v3/issues/labels/#remove-a-label-from-an-issue

replaceAllLabelsForIssueR :: Foldable f => Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> Request 'RW (Vector IssueLabel)

Replace all labels on an issue. See https://developer.github.com/v3/issues/labels/#replace-all-labels-for-an-issue

Sending an empty list will remove all labels from the issue.

removeAllLabelsFromIssueR :: Name Owner -> Name Repo -> Id Issue -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Remove all labels from an issue. See https://developer.github.com/v3/issues/labels/#remove-all-labels-from-an-issue

labelsOnMilestoneR :: forall (k :: RW). Name Owner -> Name Repo -> Id Milestone -> FetchCount -> Request k (Vector IssueLabel)

Query labels for every issue in a milestone. See https://developer.github.com/v3/issues/labels/#get-labels-for-every-issue-in-a-milestone

Milestone

See https://developer.github.com/v3/issues/milestones/

milestonesR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector Milestone)

List milestones for a repository. See https://developer.github.com/v3/issues/milestones/#list-milestones-for-a-repository

milestoneR :: forall (k :: RW). Name Owner -> Name Repo -> Id Milestone -> Request k Milestone

Query a single milestone. See https://developer.github.com/v3/issues/milestones/#get-a-single-milestone

createMilestoneR :: Name Owner -> Name Repo -> NewMilestone -> Request 'RW Milestone

Create a milestone. See https://developer.github.com/v3/issues/milestones/#create-a-milestone

updateMilestoneR :: Name Owner -> Name Repo -> Id Milestone -> UpdateMilestone -> Request 'RW Milestone

Update a milestone. See https://developer.github.com/v3/issues/milestones/#update-a-milestone

deleteMilestoneR :: Name Owner -> Name Repo -> Id Milestone -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a milestone. See https://developer.github.com/v3/issues/milestones/#delete-a-milestone

Organizations

See https://developer.github.com/v3/orgs/

Missing endpoints:

List your organizations

List all organizations

Edit an organization

publicOrganizationsForR :: forall (k :: RW). Name User -> FetchCount -> Request k (Vector SimpleOrganization)

List public user organizations. See https://developer.github.com/v3/orgs/#list-user-organizations

publicOrganizationR :: forall (k :: RW). Name Organization -> Request k Organization

Query an organization. See https://developer.github.com/v3/orgs/#get-an-organization

organizationsR :: forall (k :: RW). FetchCount -> Request k (Vector SimpleOrganization)

List all user organizations. See https://developer.github.com/v3/orgs/#list-your-organizations

Members

See https://developer.github.com/v3/orgs/members/

Missing endpoints: All except Members List and Check Membership

membersOfR :: forall (k :: RW). Name Organization -> FetchCount -> Request k (Vector SimpleUser)

All the users who are members of the specified organization.

See https://developer.github.com/v3/orgs/members/#members-list

membersOfWithR :: forall (k :: RW). Name Organization -> OrgMemberFilter -> OrgMemberRole -> FetchCount -> Request k (Vector SimpleUser)

membersOfR with filters.

See https://developer.github.com/v3/orgs/members/#members-list

isMemberOfR :: forall (rw :: RW). Name User -> Name Organization -> GenRequest ('MtStatus :: MediaType Type) rw Bool

Check if a user is a member of an organization.

See https://developer.github.com/v3/orgs/members/#check-membership

orgInvitationsR :: Name Organization -> FetchCount -> Request 'RA (Vector Invitation)

List pending organization invitations

See https://developer.github.com/v3/orgs/members/#list-pending-organization-invitations

orgMembershipR :: Name User -> Name Organization -> Request 'RA Membership

Get user membership information in an organization

See https://docs.github.com/en/rest/orgs/members#get-organization-membership-for-a-user

Outside Collaborators

See https://developer.github.com/v3/orgs/outside_collaborators/

Missing endpoints: All except Outside Collaborator List

outsideCollaboratorsR :: forall (k :: RW). Name Organization -> FetchCount -> Request k (Vector SimpleUser)

All the users who are outside collaborators of the specified organization.

See https://developer.github.com/v3/orgs/outside_collaborators/#list-outside-collaborators

Teams

See https://developer.github.com/v3/orgs/teams/

Missing endpoints:

Query team member (deprecated)

Add team member (deprecated)

Remove team member (deprecated)

Check if a team manages a repository

Add team repository

Remove team repository

teamsOfR :: forall (k :: RW). Name Organization -> FetchCount -> Request k (Vector SimpleTeam)

List teams. See https://developer.github.com/v3/orgs/teams/#list-teams

teamInfoForR :: forall (k :: RW). Id Team -> Request k Team

Query team. See https://developer.github.com/v3/orgs/teams/#get-team

createTeamForR :: Name Organization -> CreateTeam -> Request 'RW Team

Create team. See https://developer.github.com/v3/orgs/teams/#create-team

editTeamR :: Id Team -> EditTeam -> Request 'RW Team

Edit team. See https://developer.github.com/v3/orgs/teams/#edit-team

deleteTeamR :: Id Team -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

listTeamMembersR :: Id Team -> TeamMemberRole -> FetchCount -> Request 'RA (Vector SimpleUser)

List team members.

See https://developer.github.com/v3/orgs/teams/#list-team-members

listTeamReposR :: forall (k :: RW). Id Team -> FetchCount -> Request k (Vector Repo)

Query team repositories. See https://developer.github.com/v3/orgs/teams/#list-team-repos

teamMembershipInfoForR :: forall (k :: RW). Id Team -> Name Owner -> Request k TeamMembership

Query team membership. See <https://developer.github.com/v3/orgs/teams/#get-team-membership

addTeamMembershipForR :: Id Team -> Name Owner -> Role -> Request 'RW TeamMembership

Add team membership. See https://developer.github.com/v3/orgs/teams/#add-team-membership

deleteTeamMembershipForR :: Id Team -> Name Owner -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Remove team membership. See https://developer.github.com/v3/orgs/teams/#remove-team-membership

listTeamsCurrentR :: FetchCount -> Request 'RA (Vector Team)

List user teams. See https://developer.github.com/v3/orgs/teams/#list-user-teams

Pull Requests

See https://developer.github.com/v3/pulls/

pullRequestsForR :: forall (k :: RW). Name Owner -> Name Repo -> PullRequestMod -> FetchCount -> Request k (Vector SimplePullRequest)

List pull requests. See https://developer.github.com/v3/pulls/#list-pull-requests

pullRequestR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> Request k PullRequest

Query a single pull request. See https://developer.github.com/v3/pulls/#get-a-single-pull-request

pullRequestPatchR :: forall (rw :: RW). Name Owner -> Name Repo -> IssueNumber -> GenRequest ('MtPatch :: MediaType Type) rw ByteString

Query a single pull request to obtain the patch See https://developer.github.com/v3/pulls/#get-a-single-pull-request

pullRequestDiffR :: forall (rw :: RW). Name Owner -> Name Repo -> IssueNumber -> GenRequest ('MtDiff :: MediaType Type) rw ByteString

Query a single pull request to obtain the diff See https://developer.github.com/v3/pulls/#get-a-single-pull-request

createPullRequestR :: Name Owner -> Name Repo -> CreatePullRequest -> Request 'RW PullRequest

Create a pull request. See https://developer.github.com/v3/pulls/#create-a-pull-request

updatePullRequestR :: Name Owner -> Name Repo -> IssueNumber -> EditPullRequest -> Request 'RW PullRequest

Update a pull request. See https://developer.github.com/v3/pulls/#update-a-pull-request

pullRequestCommitsR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> FetchCount -> Request k (Vector Commit)

List commits on a pull request. See https://developer.github.com/v3/pulls/#list-commits-on-a-pull-request

pullRequestFilesR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> FetchCount -> Request k (Vector File)

List pull requests files. See https://developer.github.com/v3/pulls/#list-pull-requests-files

isPullRequestMergedR :: forall (rw :: RW). Name Owner -> Name Repo -> IssueNumber -> GenRequest ('MtStatus :: MediaType Type) rw Bool

Query if a pull request has been merged. See https://developer.github.com/v3/pulls/#get-if-a-pull-request-has-been-merged

mergePullRequestR :: Name Owner -> Name Repo -> IssueNumber -> Maybe Text -> GenRequest ('MtStatus :: MediaType Type) 'RW MergeResult

Merge a pull request (Merge Button). https://developer.github.com/v3/pulls/#merge-a-pull-request-merge-button

Review comments

See https://developer.github.com/v3/pulls/comments/

Missing endpoints:

List comments in a repository

Edit a comment

Delete a comment

pullRequestCommentsR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> FetchCount -> Request k (Vector Comment)

List comments on a pull request. See https://developer.github.com/v3/pulls/comments/#list-comments-on-a-pull-request

pullRequestCommentR :: forall (k :: RW). Name Owner -> Name Repo -> Id Comment -> Request k Comment

Query a single comment. See https://developer.github.com/v3/pulls/comments/#get-a-single-comment

createPullCommentR :: Name Owner -> Name Repo -> IssueNumber -> Text -> Text -> Int -> Text -> Request 'RW Comment

Create a comment.

See https://developer.github.com/v3/pulls/comments/#create-a-comment

createPullCommentReplyR :: Name Owner -> Name Repo -> IssueNumber -> Id Comment -> Text -> Request 'RW Comment

Create a comment reply.

See https://developer.github.com/v3/pulls/comments/#create-a-review-comment-reply

Pull request reviews

See https://developer.github.com/v3/pulls/reviews/

Missing endpoints:

Delete a pending review

Create a pull request review

Submit a pull request review

Dismiss a pull request review

pullRequestReviewsR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> FetchCount -> Request k (Vector Review)

List reviews for a pull request. See https://developer.github.com/v3/pulls/reviews/#list-reviews-on-a-pull-request

pullRequestReviewR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> Id Review -> Request k Review

Query a single pull request review. see https://developer.github.com/v3/pulls/reviews/#get-a-single-review

pullRequestReviewCommentsR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> Id Review -> Request k [ReviewComment]

Query the comments for a single pull request review. see https://developer.github.com/v3/pulls/reviews/#get-comments-for-a-single-review

Repositories

See https://developer.github.com/v3/repos/

Missing endpoints:

List all public repositories

List Teams

Query Branch

Enabling and disabling branch protection

currentUserReposR :: forall (k :: RW). RepoPublicity -> FetchCount -> Request k (Vector Repo)

List your repositories. See https://docs.github.com/en/rest/reference/repos#list-repositories-for-the-authenticated-user

userReposR :: forall (k :: RW). Name Owner -> RepoPublicity -> FetchCount -> Request k (Vector Repo)

List user repositories. See https://docs.github.com/en/rest/reference/repos#list-repositories-for-a-user

organizationReposR :: forall (k :: RW). Name Organization -> RepoPublicity -> FetchCount -> Request k (Vector Repo)

List organization repositories. See https://docs.github.com/en/rest/reference/repos#list-organization-repositories

repositoryR :: forall (k :: RW). Name Owner -> Name Repo -> Request k Repo

Query single repository. See https://developer.github.com/v3/repos/#get

contributorsR

Arguments

:: forall (k :: RW). Name Owner

-> Name Repo

-> Bool
Include anonymous

-> FetchCount

-> Request k (Vector Contributor)

List contributors. See https://developer.github.com/v3/repos/#list-contributors

languagesForR :: forall (k :: RW). Name Owner -> Name Repo -> Request k Languages

List languages. See https://developer.github.com/v3/repos/#list-languages

tagsForR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector Tag)

List tags. See https://developer.github.com/v3/repos/#list-tags

branchesForR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector Branch)

List branches. See https://developer.github.com/v3/repos/#list-branches

Collaborators

See https://developer.github.com/v3/repos/collaborators/

collaboratorsOnR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector SimpleUser)

List collaborators. See https://developer.github.com/v3/repos/collaborators/#list-collaborators

collaboratorPermissionOnR

Arguments

:: forall (rw :: RW). Name Owner
Repository owner

-> Name Repo
Repository name

-> Name User
Collaborator to check permissions of.

-> GenRequest ('MtJSON :: MediaType Type) rw CollaboratorWithPermission

Review a user's permission level. https://developer.github.com/v3/repos/collaborators/#review-a-users-permission-level

isCollaboratorOnR

Arguments

:: forall (rw :: RW). Name Owner
Repository owner

-> Name Repo
Repository name

-> Name User
Collaborator?

-> GenRequest ('MtStatus :: MediaType Type) rw Bool

Check if a user is a collaborator. See https://developer.github.com/v3/repos/collaborators/#check-if-a-user-is-a-collaborator

addCollaboratorR

Arguments

:: Name Owner
Repository owner

-> Name Repo
Repository name

-> Name User
Collaborator to add

-> GenRequest ('MtJSON :: MediaType Type) 'RW (Maybe RepoInvitation)

Invite a user as a collaborator. See https://developer.github.com/v3/repos/collaborators/#add-user-as-a-collaborator

Comments

See https://developer.github.com/v3/repos/comments/

Missing endpoints:

Create a commit comment

Update a commit comment

Delete a commit comment

commentsForR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector Comment)

List commit comments for a repository. See https://developer.github.com/v3/repos/comments/#list-commit-comments-for-a-repository

commitCommentsForR :: forall (k :: RW). Name Owner -> Name Repo -> Name Commit -> FetchCount -> Request k (Vector Comment)

List comments for a single commit. See https://developer.github.com/v3/repos/comments/#list-comments-for-a-single-commit

commitCommentForR :: forall (k :: RW). Name Owner -> Name Repo -> Id Comment -> Request k Comment

Query a single commit comment. See https://developer.github.com/v3/repos/comments/#get-a-single-commit-comment

Commits

See https://developer.github.com/v3/repos/commits/

commitsForR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector Commit)

List commits on a repository. See https://developer.github.com/v3/repos/commits/#list-commits-on-a-repository

commitsWithOptionsForR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> [CommitQueryOption] -> Request k (Vector Commit)

List commits on a repository. See https://developer.github.com/v3/repos/commits/#list-commits-on-a-repository

commitR :: forall (k :: RW). Name Owner -> Name Repo -> Name Commit -> Request k Commit

Query a single commit. See https://developer.github.com/v3/repos/commits/#get-a-single-commit

diffR :: forall (k :: RW). Name Owner -> Name Repo -> Name Commit -> Name Commit -> Request k Diff

Compare two commits. See https://developer.github.com/v3/repos/commits/#compare-two-commits

Reactions

See https://docs.github.com/en/rest/reactions/reactions?apiVersion=2022-11-28

issueReactionsR :: forall (k :: RW). Name Owner -> Name Repo -> Id Issue -> FetchCount -> Request k (Vector Reaction)

List reactions for an issue. See https://docs.github.com/en/rest/reactions/reactions?apiVersion=2022-11-28#list-reactions-for-an-issue

createIssueReactionR :: Name Owner -> Name Repo -> Id Issue -> ReactionContent -> Request 'RW Reaction

Create reaction for an issue comment. See https://docs.github.com/en/rest/reactions/reactions?apiVersion=2022-11-28#create-reaction-for-an-issue

deleteIssueReactionR :: Name Owner -> Name Repo -> Id Issue -> Id Reaction -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete an issue comment reaction. See https://docs.github.com/en/rest/reactions/reactions?apiVersion=2022-11-28#delete-an-issue-reaction

commentReactionsR :: forall (k :: RW). Name Owner -> Name Repo -> Id Comment -> FetchCount -> Request k (Vector Reaction)

List reactions for an issue comment. See https://docs.github.com/en/rest/reactions/reactions?apiVersion=2022-11-28#list-reactions-for-an-issue-comment

createCommentReactionR :: Name Owner -> Name Repo -> Id Comment -> ReactionContent -> Request 'RW Reaction

Create reaction for an issue comment. See https://docs.github.com/en/rest/reactions/reactions?apiVersion=2022-11-28#create-reaction-for-an-issue-comment

deleteCommentReactionR :: Name Owner -> Name Repo -> Id Comment -> Id Reaction -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete an issue comment reaction. See https://docs.github.com/en/rest/reactions/reactions?apiVersion=2022-11-28#delete-an-issue-comment-reaction

Contents

See https://developer.github.com/v3/repos/contents/

contentsForR

Arguments

:: forall (k :: RW). Name Owner

-> Name Repo

-> Text
file or directory

-> Maybe Text
Git commit

-> Request k Content

readmeForR :: forall (k :: RW). Name Owner -> Name Repo -> Request k Content

archiveForR

Arguments

:: forall (rw :: RW). Name Owner

-> Name Repo

-> ArchiveFormat
The type of archive to retrieve

-> Maybe Text
Git commit

-> GenRequest ('MtRedirect :: MediaType Type) rw URI

Get archive link. See https://developer.github.com/v3/repos/contents/#get-archive-link

createFileR :: Name Owner -> Name Repo -> CreateFile -> Request 'RW ContentResult

Create a file. See https://developer.github.com/v3/repos/contents/#create-a-file

updateFileR :: Name Owner -> Name Repo -> UpdateFile -> Request 'RW ContentResult

Update a file. See https://developer.github.com/v3/repos/contents/#update-a-file

deleteFileR :: Name Owner -> Name Repo -> DeleteFile -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a file. See https://developer.github.com/v3/repos/contents/#delete-a-file

Deploy Keys

See https://developer.github.com/v3/repos/keys/

deployKeysForR :: Name Owner -> Name Repo -> FetchCount -> Request 'RA (Vector RepoDeployKey)

Querying deploy keys. See https://developer.github.com/v3/repos/keys/#list-deploy-keys

deployKeyForR :: Name Owner -> Name Repo -> Id RepoDeployKey -> Request 'RA RepoDeployKey

Querying a deploy key. See https://developer.github.com/v3/repos/keys/#get-a-deploy-key

createRepoDeployKeyR :: Name Owner -> Name Repo -> NewRepoDeployKey -> Request 'RW RepoDeployKey

Create a deploy key. See https://developer.github.com/v3/repos/keys/#add-a-new-deploy-key.

deleteRepoDeployKeyR :: Name Owner -> Name Repo -> Id RepoDeployKey -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a deploy key. See https://developer.github.com/v3/repos/keys/#remove-a-deploy-key

Deployments

See https://developer.github.com/v3/repos/deployments/#deployments

Missing endpoints: * Get a single deployment * Update a deployment * Get a single deployment status

deploymentsWithOptionsForR :: FromJSON a => Name Owner -> Name Repo -> FetchCount -> [DeploymentQueryOption] -> Request 'RA (Vector (Deployment a))

List deployments. See https://developer.github.com/v3/repos/deployments/#list-deployments

createDeploymentR :: (ToJSON a, FromJSON a) => Name Owner -> Name Repo -> CreateDeployment a -> Request 'RW (Deployment a)

Create a deployment. See https://developer.github.com/v3/repos/deployments/#create-a-deployment

deploymentStatusesForR :: Name Owner -> Name Repo -> Id (Deployment a) -> FetchCount -> Request 'RA (Vector DeploymentStatus)

List deployment statuses. See https://developer.github.com/v3/repos/deployments/#list-deployment-statuses

createDeploymentStatusR :: Name Owner -> Name Repo -> Id (Deployment a) -> CreateDeploymentStatus -> Request 'RW DeploymentStatus

Create a deployment status. See https://developer.github.com/v3/repos/deployments/#list-deployment-statuses

Forks

See https://developer.github.com/v3/repos/forks/

Missing endpoints:

Create a fork

forksForR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector Repo)

List forks. See https://developer.github.com/v3/repos/forks/#list-forks

Statuses

See https://developer.github.com/v3/repos/statuses/

createStatusR :: Name Owner -> Name Repo -> Name Commit -> NewStatus -> Request 'RW Status

Create a new status See https://developer.github.com/v3/repos/statuses/#create-a-status

statusesForR :: Name Owner -> Name Repo -> Name Commit -> FetchCount -> Request 'RW (Vector Status)

All statuses for a commit See https://developer.github.com/v3/repos/statuses/#list-statuses-for-a-specific-ref

statusForR :: Name Owner -> Name Repo -> Name Commit -> Request 'RW CombinedStatus

The combined status for a specific commit See https://developer.github.com/v3/repos/statuses/#get-the-combined-status-for-a-specific-ref

Webhooks

See https://developer.github.com/v3/repos/hooks/

webhooksForR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector RepoWebhook)

List hooks. See https://developer.github.com/v3/repos/hooks/#list-hooks

webhookForR :: forall (k :: RW). Name Owner -> Name Repo -> Id RepoWebhook -> Request k RepoWebhook

createRepoWebhookR :: Name Owner -> Name Repo -> NewRepoWebhook -> Request 'RW RepoWebhook

Create a hook. See https://developer.github.com/v3/repos/hooks/#create-a-hook

editRepoWebhookR :: Name Owner -> Name Repo -> Id RepoWebhook -> EditRepoWebhook -> Request 'RW RepoWebhook

Edit a hook. See https://developer.github.com/v3/repos/hooks/#edit-a-hook

testPushRepoWebhookR :: Name Owner -> Name Repo -> Id RepoWebhook -> GenRequest ('MtStatus :: MediaType Type) 'RW Bool

Test a push hook. See https://developer.github.com/v3/repos/hooks/#test-a-push-hook

pingRepoWebhookR :: Name Owner -> Name Repo -> Id RepoWebhook -> GenRequest ('MtStatus :: MediaType Type) 'RW Bool

Ping a hook. See https://developer.github.com/v3/repos/hooks/#ping-a-hook

deleteRepoWebhookR :: Name Owner -> Name Repo -> Id RepoWebhook -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a hook. See https://developer.github.com/v3/repos/hooks/#delete-a-hook

Releases

releasesR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector Release)

List releases for a repository. See https://developer.github.com/v3/repos/releases/#list-releases-for-a-repository

releaseR :: forall (k :: RW). Name Owner -> Name Repo -> Id Release -> Request k Release

Get a single release. See https://developer.github.com/v3/repos/releases/#get-a-single-release

latestReleaseR :: forall (k :: RW). Name Owner -> Name Repo -> Request k Release

Get the latest release. See https://developer.github.com/v3/repos/releases/#get-the-latest-release

releaseByTagNameR :: forall (k :: RW). Name Owner -> Name Repo -> Text -> Request k Release

Get a release by tag name See https://developer.github.com/v3/repos/releases/#get-a-release-by-tag-name

Invitations

See https://developer.github.com/v3/repos/invitations/ Missing endpoints:

Delete a repository invitation

Update a repository invitation

Decline a repository invitation

listInvitationsOnR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) k (Vector RepoInvitation)

List open invitations of a repository See https://developer.github.com/v3/repos/invitations/#list-invitations-for-a-repository

acceptInvitationFromR :: Id RepoInvitation -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Accept a repository invitation See https://developer.github.com/v3/repos/invitations/#accept-a-repository-invitation

listInvitationsForR :: forall (k :: RW). FetchCount -> Request k (Vector RepoInvitation)

List a user's repository invitations See https://developer.github.com/v3/repos/invitations/#list-a-users-repository-invitations

Search

See https://developer.github.com/v3/search/

searchReposR :: forall (k :: RW). Text -> FetchCount -> Request k (SearchResult Repo)

Search repositories. See https://developer.github.com/v3/search/#search-repositories

searchCodeR :: forall (k :: RW). Text -> FetchCount -> Request k (SearchResult Code)

Search code. See https://developer.github.com/v3/search/#search-code

searchIssuesR :: forall (k :: RW). Text -> FetchCount -> Request k (SearchResult Issue)

Search issues. See https://developer.github.com/v3/search/#search-issues

searchUsersR :: forall (k :: RW). Text -> FetchCount -> Request k (SearchResult SimpleUser)

Search users. See https://developer.github.com/v3/search/#search-code

Users

See https://developer.github.com/v3/users/

Missing endpoints:

Update the authenticated user

Query all users

userInfoForR :: forall (k :: RW). Name User -> Request k User

Query a single user. See https://developer.github.com/v3/users/#get-a-single-user

>>> github' userInfoForR "mike-burns"

or

>>> github userInfoForR (OAuth "github-token") "mike-burns"

ownerInfoForR :: forall (k :: RW). Name Owner -> Request k Owner

Query a single user or an organization. See https://developer.github.com/v3/users/#get-a-single-user

userInfoCurrentR :: Request 'RA User

Query the authenticated user. See https://developer.github.com/v3/users/#get-the-authenticated-user

Emails

See https://developer.github.com/v3/users/emails/

Missing endpoints:

Add email address(es)

Delete email address(es)

Toggle primary email visibility

currentUserEmailsR :: FetchCount -> Request 'RA (Vector Email)

List email addresses. See https://developer.github.com/v3/users/emails/#list-email-addresses-for-a-user

currentUserPublicEmailsR :: FetchCount -> Request 'RA (Vector Email)

List public email addresses. See https://developer.github.com/v3/users/emails/#list-public-email-addresses-for-a-user

Followers

See https://developer.github.com/v3/users/followers/

Missing endpoints:

Check if you are following a user

Check if one user follows another

Follow a user

Unfollow a user

usersFollowingR :: forall (k :: RW). Name User -> FetchCount -> Request k (Vector SimpleUser)

List followers of a user. See https://developer.github.com/v3/users/followers/#list-followers-of-a-user

usersFollowedByR :: forall (k :: RW). Name User -> FetchCount -> Request k (Vector SimpleUser)

List users followed by another user. See https://developer.github.com/v3/users/followers/#list-users-followed-by-another-user

Git SSH Keys

See https://developer.github.com/v3/users/keys/

publicSSHKeysR :: Request 'RA (Vector PublicSSHKey)

Querying the authenticated users' public SSH keys See https://developer.github.com/v3/users/keys/#list-your-public-keys

publicSSHKeysForR :: Name Owner -> FetchCount -> Request 'RO (Vector PublicSSHKeyBasic)

Querying public SSH keys. See https://developer.github.com/v3/users/keys/#list-public-keys-for-a-user

publicSSHKeyR :: Id PublicSSHKey -> Request 'RA PublicSSHKey

Querying a public SSH key. See https://developer.github.com/v3/users/keys/#get-a-single-public-key

createUserPublicSSHKeyR :: NewPublicSSHKey -> Request 'RW PublicSSHKey

Create a public SSH key. See https://developer.github.com/v3/users/keys/#create-a-public-key.

deleteUserPublicSSHKeyR :: Id PublicSSHKey -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a public SSH key. See https://developer.github.com/v3/users/keys/#delete-a-public-key

Rate Limit

See https://developer.github.com/v3/rate_limit/

rateLimitR :: forall (k :: RW). Request k RateLimit

Get your current rate limit status. https://developer.github.com/v3/rate_limit/#get-your-current-rate-limit-status

Actions - artifacts

See https://docs.github.com/en/rest/actions/artifacts

artifactsForR :: Name Owner -> Name Repo -> ArtifactMod -> FetchCount -> Request 'RA (WithTotalCount Artifact)

List artifacts for repository. See https://docs.github.com/en/rest/reference/actions#list-artifacts-for-a-repository

artifactR :: Name Owner -> Name Repo -> Id Artifact -> Request 'RA Artifact

Get an artifact. See https://docs.github.com/en/rest/reference/actions#get-an-artifact

deleteArtifactR :: Name Owner -> Name Repo -> Id Comment -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete an artifact. See https://docs.github.com/en/rest/reference/actions#delete-an-artifact

downloadArtifactR :: Name Owner -> Name Repo -> Id Artifact -> GenRequest ('MtRedirect :: MediaType Type) 'RW URI

Download an artifact. See https://docs.github.com/en/rest/reference/actions#download-an-artifact

artifactsForWorkflowRunR :: Name Owner -> Name Repo -> Id WorkflowRun -> FetchCount -> Request 'RA (WithTotalCount Artifact)

List artifacts for a workflow run. See https://docs.github.com/en/rest/reference/actions#list-workflow-run-artifacts

Actions - cache

See https://docs.github.com/en/rest/actions/cache

cacheUsageOrganizationR :: Name Organization -> GenRequest ('MtJSON :: MediaType Type) 'RA OrganizationCacheUsage

Get Actions cache usage for the organization. See https://docs.github.com/en/rest/actions/cache#get-github-actions-cache-usage-for-an-organization

cacheUsageByRepositoryR :: Name Organization -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount RepositoryCacheUsage)

List repositories with GitHub Actions cache usage for an organization. See https://docs.github.com/en/rest/actions/cache#list-repositories-with-github-actions-cache-usage-for-an-organization

cacheUsageR :: forall (k :: RW). Name Owner -> Name Repo -> Request k RepositoryCacheUsage

Get GitHub Actions cache usage for a repository. See https://docs.github.com/en/rest/actions/cache#get-github-actions-cache-usage-for-a-repository

cachesForRepoR :: Name Owner -> Name Repo -> CacheMod -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount Cache)

List the GitHub Actions caches for a repository. See https://docs.github.com/en/rest/actions/cache#list-github-actions-caches-for-a-repository

deleteCacheR :: Name Owner -> Name Repo -> Id Cache -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete GitHub Actions cache for a repository. See https://docs.github.com/en/rest/actions/cache#delete-a-github-actions-cache-for-a-repository-using-a-cache-id

Actions - secrets

See https://docs.github.com/en/rest/actions/secrets

organizationSecretsR :: Name Organization -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount OrganizationSecret)

List organization secrets. See https://docs.github.com/en/rest/actions/secrets#list-organization-secrets

organizationPublicKeyR :: Name Organization -> GenRequest ('MtJSON :: MediaType Type) 'RA PublicKey

List organization secrets. See https://docs.github.com/en/rest/actions/secrets#get-an-organization-public-key

organizationSecretR :: Name Organization -> Name OrganizationSecret -> GenRequest ('MtJSON :: MediaType Type) 'RA OrganizationSecret

Get an organization secret. See https://docs.github.com/en/rest/actions/secrets#get-an-organization-secret

setOrganizationSecretR :: Name Organization -> Name OrganizationSecret -> SetSecret -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Create or update an organization secret. See https://docs.github.com/en/rest/actions/secrets#create-or-update-an-organization-secret

deleteOrganizationSecretR :: Name Organization -> Name OrganizationSecret -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete an organization secret. See https://docs.github.com/en/rest/actions/secrets#delete-an-organization-secret

organizationSelectedRepositoriesForSecretR :: Name Organization -> Name OrganizationSecret -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount SelectedRepo)

Get selected repositories for an organization secret. See https://docs.github.com/en/rest/actions/secrets#list-selected-repositories-for-an-organization-secret

setOrganizationSelectedRepositoriesForSecretR :: Name Organization -> Name OrganizationSecret -> SetSelectedRepositories -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Set selected repositories for an organization secret. See https://docs.github.com/en/rest/actions/secrets#set-selected-repositories-for-an-organization-secret

addOrganizationSelectedRepositoriesForSecretR :: Name Organization -> Name OrganizationSecret -> Id Repo -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Add selected repository to an organization secret. See https://docs.github.com/en/rest/actions/secrets#add-selected-repository-to-an-organization-secret

removeOrganizationSelectedRepositoriesForSecretR :: Name Organization -> Name OrganizationSecret -> Id Repo -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Remove selected repository from an organization secret. See https://docs.github.com/en/rest/actions/secrets#remove-selected-repository-from-an-organization-secret

repoSecretsR :: Name Owner -> Name Repo -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount RepoSecret)

List repository secrets. See https://docs.github.com/en/rest/actions/secrets#list-repository-secrets

repoPublicKeyR :: Name Owner -> Name Organization -> GenRequest ('MtJSON :: MediaType Type) 'RA PublicKey

Get a repository public key. See https://docs.github.com/en/rest/actions/secrets#get-a-repository-public-key

repoSecretR :: Name Owner -> Name Organization -> Name RepoSecret -> GenRequest ('MtJSON :: MediaType Type) 'RA RepoSecret

Get a repository secret. See https://docs.github.com/en/rest/actions/secrets#get-a-repository-secret

setRepoSecretR :: Name Owner -> Name Organization -> Name RepoSecret -> SetRepoSecret -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Create or update a repository secret. See https://docs.github.com/en/rest/actions/secrets#create-or-update-a-repository-secret

deleteRepoSecretR :: Name Owner -> Name Organization -> Name RepoSecret -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a repository secret. See https://docs.github.com/en/rest/actions/secrets#delete-a-repository-secret

environmentSecretsR :: Id Repo -> Name Environment -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount RepoSecret)

List environment secrets. See https://docs.github.com/en/rest/actions/secrets#list-environment-secrets

environmentPublicKeyR :: Id Repo -> Name Environment -> GenRequest ('MtJSON :: MediaType Type) 'RA PublicKey

Get an environment public key. See https://docs.github.com/en/rest/actions/secrets#get-an-environment-public-key

environmentSecretR :: Id Repo -> Name Environment -> Name RepoSecret -> GenRequest ('MtJSON :: MediaType Type) 'RA RepoSecret

Get an environment secret See https://docs.github.com/en/rest/actions/secrets#get-an-environment-secret

setEnvironmentSecretR :: Id Repo -> Name Environment -> Name RepoSecret -> SetRepoSecret -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Create or update an environment secret. See https://docs.github.com/en/rest/actions/secrets#create-or-update-an-environment-secret

deleteEnvironmentSecretR :: Id Repo -> Name Environment -> Name RepoSecret -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete an environment secret. See https://docs.github.com/en/rest/actions/secrets#delete-an-environment-secret

Actions - workflow jobs

See https://docs.github.com/en/rest/actions/workflow-jobs

jobR :: Name Owner -> Name Repo -> Id Job -> Request 'RA Job

Get a job for a workflow run. See https://docs.github.com/en/rest/actions/workflow-jobs#get-a-job-for-a-workflow-run

downloadJobLogsR :: Name Owner -> Name Repo -> Id Job -> GenRequest ('MtRedirect :: MediaType Type) 'RO URI

Download job logs for a workflow run. See https://docs.github.com/en/rest/actions/workflow-jobs#download-job-logs-for-a-workflow-run

jobsForWorkflowRunAttemptR :: Name Owner -> Name Repo -> Id WorkflowRun -> Id RunAttempt -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount Job)

List jobs for a workflow run attempt. See https://docs.github.com/en/rest/actions/workflow-jobs#list-jobs-for-a-workflow-run-attempt

jobsForWorkflowRunR :: Name Owner -> Name Repo -> Id WorkflowRun -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount Job)

List jobs for a workflow run. See https://docs.github.com/en/rest/actions/workflow-jobs#list-jobs-for-a-workflow-run

Actions - workflow runs

See https://docs.github.com/en/rest/actions/workflow-runs

reRunJobR :: Name Owner -> Name Repo -> Id Job -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Re-run a job from a workflow run. See https://docs.github.com/en/rest/actions/workflow-runs#re-run-a-job-from-a-workflow-run

workflowRunsR :: Name Owner -> Name Repo -> WorkflowRunMod -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount WorkflowRun)

List workflow runs for a repository. See https://docs.github.com/en/rest/actions/workflow-runs#list-workflow-runs-for-a-repository

workflowRunR :: Name Owner -> Name Repo -> Id WorkflowRun -> GenRequest ('MtJSON :: MediaType Type) 'RA WorkflowRun

Get a workflow run. See https://docs.github.com/en/rest/actions/workflow-runs#get-a-workflow-run

deleteWorkflowRunR :: Name Owner -> Name Repo -> Id WorkflowRun -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a workflow run. See https://docs.github.com/en/rest/actions/workflow-runs#delete-a-workflow-run

workflowRunReviewHistoryR :: Name Owner -> Name Repo -> Id WorkflowRun -> GenRequest ('MtJSON :: MediaType Type) 'RA (Vector ReviewHistory)

Get the review history for a workflow run. See https://docs.github.com/en/rest/actions/workflow-runs#get-the-review-history-for-a-workflow-run

approveWorkflowRunR :: Name Owner -> Name Repo -> Id WorkflowRun -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Approve a workflow run for a fork pull request. See https://docs.github.com/en/rest/actions/workflow-runs#approve-a-workflow-run-for-a-fork-pull-request

workflowRunAttemptR :: Name Owner -> Name Repo -> Id WorkflowRun -> Id RunAttempt -> GenRequest ('MtJSON :: MediaType Type) 'RA WorkflowRun

Get a workflow run attempt. See https://docs.github.com/en/rest/actions/workflow-runs#get-a-workflow-run-attempt

downloadWorkflowRunAttemptLogsR :: Name Owner -> Name Repo -> Id WorkflowRun -> Id RunAttempt -> GenRequest ('MtRedirect :: MediaType Type) 'RO URI

Download workflow run attempt logs. See https://docs.github.com/en/rest/actions/workflow-runs#download-workflow-run-attempt-logs

cancelWorkflowRunR :: Name Owner -> Name Repo -> Id WorkflowRun -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Cancel a workflow run. See https://docs.github.com/en/rest/actions/workflow-runs#cancel-a-workflow-run

downloadWorkflowRunLogsR :: Name Owner -> Name Repo -> Id WorkflowRun -> GenRequest ('MtRedirect :: MediaType Type) 'RA URI

Download workflow run logs. See https://docs.github.com/en/rest/actions/workflow-runs#download-workflow-run-logs

deleteWorkflowRunLogsR :: Name Owner -> Name Repo -> Id WorkflowRun -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete workflow run logs. See https://docs.github.com/en/rest/actions/workflow-runs#delete-workflow-run-logs

reRunWorkflowR :: Name Owner -> Name Repo -> Id WorkflowRun -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Re-run a workflow. See https://docs.github.com/en/rest/actions/workflow-runs#re-run-a-workflow

reRunFailedJobsR :: Name Owner -> Name Repo -> Id WorkflowRun -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Re-run failed jobs from a workflow run. See https://docs.github.com/en/rest/actions/re-run-failed-jobs-from-a-workflow-run

workflowRunsForWorkflowR :: IsPathPart idOrName => Name Owner -> Name Repo -> idOrName -> WorkflowRunMod -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount WorkflowRun)

List workflow runs for a workflow. See https://docs.github.com/en/rest/actions/workflow-runs#list-workflow-runs-for-a-workflow

Actions - workflows

See https://docs.github.com/en/rest/actions/workflows

repositoryWorkflowsR :: Name Owner -> Name Repo -> FetchCount -> GenRequest ('MtJSON :: MediaType Type) 'RA (WithTotalCount Workflow)

List repository workflows. See https://docs.github.com/en/rest/actions/workflows#list-repository-workflows

workflowR :: IsPathPart idOrName => Name Owner -> Name Repo -> idOrName -> GenRequest ('MtJSON :: MediaType Type) 'RA Workflow

Get a workflow. See https://docs.github.com/en/rest/actions/workflows#get-a-workflow

disableWorkflowR :: IsPathPart idOrName => Name Owner -> Name Repo -> idOrName -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Disable a workflow. See https://docs.github.com/en/rest/actions/workflows#disable-a-workflow

triggerWorkflowR :: (ToJSON a, IsPathPart idOrName) => Name Owner -> Name Repo -> idOrName -> CreateWorkflowDispatchEvent a -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Create a workflow dispatch event. See https://docs.github.com/en/rest/actions/workflows#create-a-workflow-dispatch-event

enableWorkflowR :: IsPathPart idOrName => Name Owner -> Name Repo -> idOrName -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Enable a workflow. See https://docs.github.com/en/rest/actions/workflows#enable-a-workflow

Data definitions

module GitHub.Data

Request handling

module GitHub.Request