GitHub.Data.Deployments

Documentation

data DeploymentQueryOption

Constructors

DeploymentQuerySha !Text

DeploymentQueryRef !Text

DeploymentQueryTask !Text

DeploymentQueryEnvironment !Text

renderDeploymentQueryOption :: DeploymentQueryOption -> (ByteString, ByteString)

data Deployment a

Constructors

Deployment

Fields

deploymentUrl :: !URL

deploymentId :: !(Id (Deployment a))

deploymentSha :: !(Name (Deployment a))

deploymentRef :: !Text

deploymentTask :: !Text

deploymentPayload :: !(Maybe a)

deploymentEnvironment :: !Text

deploymentDescription :: !Text

deploymentCreator :: !SimpleUser

deploymentCreatedAt :: !UTCTime

deploymentUpdatedAt :: !UTCTime

deploymentStatusesUrl :: !URL

deploymentRepositoryUrl :: !URL

data CreateDeployment a

Constructors

CreateDeployment

Fields

createDeploymentRef :: !Text

Required. The ref to deploy. This can be a branch, tag, or SHA.

createDeploymentTask :: !(Maybe Text)

Specifies a task to execute (e.g., deploy or deploy:migrations). Default: deploy

createDeploymentAutoMerge :: !(Maybe Bool)

Attempts to automatically merge the default branch into the requested ref, if it is behind the default branch. Default: true

createDeploymentRequiredContexts :: !(Maybe (Vector Text))

The status contexts to verify against commit status checks. If this parameter is omitted, then all unique contexts will be verified before a deployment is created. To bypass checking entirely pass an empty array. Defaults to all unique contexts.

createDeploymentPayload :: !(Maybe a)

JSON payload with extra information about the deployment. Default: ""

createDeploymentEnvironment :: !(Maybe Text)

Name for the target deployment environment (e.g., production, staging, qa). Default: production

createDeploymentDescription :: !(Maybe Text)

Short description of the deployment. Default: ""

data DeploymentStatus

Constructors

DeploymentStatus

Fields

deploymentStatusUrl :: !URL

deploymentStatusId :: !(Id DeploymentStatus)

deploymentStatusState :: !DeploymentStatusState

deploymentStatusCreator :: !SimpleUser

deploymentStatusDescription :: !Text

deploymentStatusTargetUrl :: !URL

deploymentStatusCreatedAt :: !UTCTime

deploymentStatusUpdatedAt :: !UTCTime

deploymentStatusDeploymentUrl :: !URL

deploymentStatusRepositoryUrl :: !URL

data DeploymentStatusState

Constructors

DeploymentStatusError

DeploymentStatusFailure

DeploymentStatusPending

DeploymentStatusSuccess

DeploymentStatusInactive

data CreateDeploymentStatus

Constructors

CreateDeploymentStatus

Fields

createDeploymentStatusState :: !DeploymentStatusState

Required. The state of the status. Can be one of error, failure, pending, or success.

createDeploymentStatusTargetUrl :: !(Maybe Text)

The target URL to associate with this status. This URL should contain output to keep the user updated while the task is running or serve as historical information for what happened in the deployment. Default: ""

createDeploymentStatusDescription :: !(Maybe Text)

A short description of the status. Maximum length of 140 characters. Default: ""