GitHub.Data.Issues

Documentation

data Issue

Constructors

Issue

Fields

issueClosedAt :: !(Maybe UTCTime)

issueUpdatedAt :: !UTCTime

issueEventsUrl :: !URL

issueHtmlUrl :: !(Maybe URL)

issueClosedBy :: !(Maybe SimpleUser)

issueLabels :: !(Vector IssueLabel)

issueNumber :: !IssueNumber

issueAssignees :: !(Vector SimpleUser)

issueUser :: !SimpleUser

issueTitle :: !Text

issuePullRequest :: !(Maybe PullRequestReference)

issueUrl :: !URL

issueCreatedAt :: !UTCTime

issueBody :: !(Maybe Text)

issueState :: !IssueState

issueId :: !(Id Issue)

issueComments :: !Int

issueMilestone :: !(Maybe Milestone)

issueStateReason :: !(Maybe IssueStateReason)

data NewIssue

Constructors

NewIssue

Fields

newIssueTitle :: !Text

newIssueBody :: !(Maybe Text)

newIssueAssignees :: !(Vector (Name User))

newIssueMilestone :: !(Maybe (Id Milestone))

newIssueLabels :: !(Maybe (Vector (Name IssueLabel)))

data EditIssue

Constructors

EditIssue

Fields

editIssueTitle :: !(Maybe Text)

editIssueBody :: !(Maybe Text)

editIssueAssignees :: !(Maybe (Vector (Name User)))

editIssueState :: !(Maybe IssueState)

editIssueMilestone :: !(Maybe (Id Milestone))

editIssueLabels :: !(Maybe (Vector (Name IssueLabel)))

data IssueComment

Constructors

IssueComment

Fields

issueCommentUpdatedAt :: !UTCTime

issueCommentUser :: !SimpleUser

issueCommentUrl :: !URL

issueCommentHtmlUrl :: !URL

issueCommentCreatedAt :: !UTCTime

issueCommentBody :: !Text

issueCommentId :: !Int

data EventType

See https://developer.github.com/v3/issues/events/#events-1

Constructors

Mentioned
The actor was @mentioned in an issue body.

Subscribed
The actor subscribed to receive notifications for an issue.

Unsubscribed
The issue was unsubscribed from by the actor.

Referenced
The issue was referenced from a commit message. The commit_id attribute is the commit SHA1 of where that happened.

Merged
The issue was merged by the actor. The commit_id attribute is the SHA1 of the HEAD commit that was merged.

Assigned
The issue was assigned to the actor.

Closed
The issue was closed by the actor. When the commit_id is present, it identifies the commit that closed the issue using “closes / fixes #NN” syntax.

Reopened
The issue was reopened by the actor.

ActorUnassigned
The issue was unassigned to the actor

Labeled
A label was added to the issue.

Unlabeled
A label was removed from the issue.

Milestoned
The issue was added to a milestone.

Demilestoned
The issue was removed from a milestone.

Renamed
The issue title was changed.

Locked
The issue was locked by the actor.

Unlocked
The issue was unlocked by the actor.

HeadRefDeleted
The pull request’s branch was deleted.

HeadRefRestored
The pull request’s branch was restored.

ReviewRequested
The actor requested review from the subject on this pull request.

ReviewDismissed
The actor dismissed a review from the pull request.

ReviewRequestRemoved
The actor removed the review request for the subject on this pull request.

MarkedAsDuplicate
A user with write permissions marked an issue as a duplicate of another issue or a pull request as a duplicate of another pull request.

UnmarkedAsDuplicate
An issue that a user had previously marked as a duplicate of another issue is no longer considered a duplicate, or a pull request that a user had previously marked as a duplicate of another pull request is no longer considered a duplicate.

AddedToProject
The issue was added to a project board.

MovedColumnsInProject
The issue was moved between columns in a project board.

RemovedFromProject
The issue was removed from a project board.

ConvertedNoteToIssue
The issue was created by converting a note in a project board to an issue.

data IssueEvent

Issue event

Constructors

IssueEvent

Fields

issueEventActor :: !SimpleUser

issueEventType :: !EventType

issueEventCommitId :: !(Maybe Text)

issueEventUrl :: !URL

issueEventCreatedAt :: !UTCTime

issueEventId :: !Int

issueEventIssue :: !(Maybe Issue)

issueEventLabel :: !(Maybe IssueLabel)