GitHub.Data.Options

Description

Module with modifiers for pull requests' and issues' listings.

Common modifiers

stateOpen :: HasState mod => mod

stateClosed :: HasState mod => mod

stateAll :: HasState mod => mod

sortAscending :: HasDirection mod => mod

sortDescending :: HasDirection mod => mod

sortByCreated :: HasCreatedUpdated mod => mod

sortByUpdated :: HasCreatedUpdated mod => mod

Pull Requests

data PullRequestMod

See https://developer.github.com/v3/pulls/#parameters.

prModToQueryString :: PullRequestMod -> QueryString

optionsBase :: Text -> PullRequestMod

optionsNoBase :: PullRequestMod

optionsHead :: Text -> PullRequestMod

optionsNoHead :: PullRequestMod

sortByPopularity :: PullRequestMod

sortByLongRunning :: PullRequestMod

Issues

data IssueMod

See https://docs.github.com/en/rest/reference/issues#list-issues-assigned-to-the-authenticated-user--parameters.

issueModToQueryString :: IssueMod -> QueryString

sortByComments :: HasComments mod => mod

optionsLabels :: (HasLabels mod, Foldable f) => f (Name IssueLabel) -> mod

optionsSince :: HasSince mod => UTCTime -> mod

optionsSinceAll :: HasSince mod => mod

optionsAssignedIssues :: IssueMod

optionsCreatedIssues :: IssueMod

optionsMentionedIssues :: IssueMod

optionsSubscribedIssues :: IssueMod

optionsAllIssues :: IssueMod

Repo issues

data IssueRepoMod

See https://developer.github.com/v3/issues/#parameters-1.

issueRepoModToQueryString :: IssueRepoMod -> QueryString

optionsCreator :: Name User -> IssueRepoMod

Issues created by a certain user.

optionsMentioned :: Name User -> IssueRepoMod

Issue mentioning the given user.

optionsIrrelevantMilestone :: IssueRepoMod

Don't care about milestones (default).

optionsAnyMilestone means there should be some milestone, but it can be any.

See https://developer.github.com/v3/issues/#list-issues-for-a-repository

optionsAnyMilestone :: IssueRepoMod

Issues that have a milestone.

optionsNoMilestone :: IssueRepoMod

Issues that have no milestone.

optionsMilestone :: Id Milestone -> IssueRepoMod

Issues with the given milestone.

optionsIrrelevantAssignee :: IssueRepoMod

Issues with or without assignee (default).

optionsAnyAssignee :: IssueRepoMod

Issues assigned to someone.

optionsNoAssignee :: IssueRepoMod

Issues assigned to nobody.

optionsAssignee :: Name User -> IssueRepoMod

Issues assigned to a specific user.

Actions artifacts

data ArtifactMod

See https://docs.github.com/en/rest/actions/artifacts#list-artifacts-for-a-repository.

artifactModToQueryString :: ArtifactMod -> QueryString

optionsArtifactName :: Text -> ArtifactMod

Filters artifacts by exact match on their name field.

Actions cache

data CacheMod

See https://docs.github.com/en/rest/actions/cache#list-github-actions-caches-for-a-repository.

cacheModToQueryString :: CacheMod -> QueryString

optionsRef :: Text -> CacheMod

optionsNoRef :: CacheMod

optionsKey :: Text -> CacheMod

optionsNoKey :: CacheMod

optionsDirectionAsc :: CacheMod

optionsDirectionDesc :: CacheMod

sortByCreatedAt :: CacheMod

sortByLastAccessedAt :: CacheMod

sortBySizeInBytes :: CacheMod

Actions workflow runs

data WorkflowRunMod

See https://docs.github.com/en/rest/actions/workflow-runs#list-workflow-runs-for-a-repository.

workflowRunModToQueryString :: WorkflowRunMod -> QueryString

optionsWorkflowRunActor :: Text -> WorkflowRunMod

optionsWorkflowRunBranch :: Text -> WorkflowRunMod

optionsWorkflowRunEvent :: Text -> WorkflowRunMod

optionsWorkflowRunStatus :: Text -> WorkflowRunMod

optionsWorkflowRunCreated :: Text -> WorkflowRunMod

optionsWorkflowRunHeadSha :: Text -> WorkflowRunMod

Data

data IssueState

Issue or PullRequest state

Constructors

StateOpen

StateClosed

data IssueStateReason

Issue state reason

Constructors

StateReasonCompleted

StateReasonDuplicate

StateReasonNotPlanned

StateReasonReopened

data MergeableState

PullRequest mergeable_state

Constructors

StateUnknown

StateClean

StateDirty

StateUnstable

StateBlocked

StateBehind

StateDraft

Internal

class HasState mod

Minimal complete definition

state

class HasDirection mod

Minimal complete definition

sortDir

class HasCreatedUpdated mod

Minimal complete definition

sortByCreated, sortByUpdated

class HasComments mod

Minimal complete definition

sortByComments

class HasLabels mod

Minimal complete definition

optionsLabels

class HasSince mod

Minimal complete definition

optionsSince, optionsSinceAll