GitHub.Data.Repos

Description

This module also exports FromJSON a => FromJSON (HashMap Language a) orphan-ish instance for aeson < 1

Documentation

data Repo

Constructors

Repo

Fields

repoId :: !(Id Repo)

repoName :: !(Name Repo)

repoOwner :: !SimpleOwner

repoPrivate :: !Bool

repoHtmlUrl :: !URL

repoDescription :: !(Maybe Text)

repoFork :: !(Maybe Bool)

repoUrl :: !URL

repoGitUrl :: !(Maybe URL)

repoSshUrl :: !(Maybe URL)

repoCloneUrl :: !(Maybe URL)

repoHooksUrl :: !URL

repoSvnUrl :: !(Maybe URL)

repoHomepage :: !(Maybe Text)

repoLanguage :: !(Maybe Language)

repoForksCount :: !Int

repoStargazersCount :: !Int

repoWatchersCount :: !Int

repoSize :: !(Maybe Int)

repoDefaultBranch :: !(Maybe Text)

repoOpenIssuesCount :: !Int

repoHasIssues :: !(Maybe Bool)

repoHasProjects :: !(Maybe Bool)

repoHasWiki :: !(Maybe Bool)

repoHasPages :: !(Maybe Bool)

repoHasDownloads :: !(Maybe Bool)

repoArchived :: !Bool

repoDisabled :: !Bool

repoPushedAt :: !(Maybe UTCTime)

this is Nothing for new repositories

repoCreatedAt :: !(Maybe UTCTime)

repoUpdatedAt :: !(Maybe UTCTime)

repoPermissions :: !(Maybe RepoPermissions)

Repository permissions as they relate to the authenticated user.

data CodeSearchRepo

Constructors

CodeSearchRepo

Fields

codeSearchRepoId :: !(Id Repo)

codeSearchRepoName :: !(Name Repo)

codeSearchRepoOwner :: !SimpleOwner

codeSearchRepoPrivate :: !Bool

codeSearchRepoHtmlUrl :: !URL

codeSearchRepoDescription :: !(Maybe Text)

codeSearchRepoFork :: !(Maybe Bool)

codeSearchRepoUrl :: !URL

codeSearchRepoGitUrl :: !(Maybe URL)

codeSearchRepoSshUrl :: !(Maybe URL)

codeSearchRepoCloneUrl :: !(Maybe URL)

codeSearchRepoHooksUrl :: !URL

codeSearchRepoSvnUrl :: !(Maybe URL)

codeSearchRepoHomepage :: !(Maybe Text)

codeSearchRepoLanguage :: !(Maybe Language)

codeSearchRepoSize :: !(Maybe Int)

codeSearchRepoDefaultBranch :: !(Maybe Text)

codeSearchRepoHasIssues :: !(Maybe Bool)

codeSearchRepoHasProjects :: !(Maybe Bool)

codeSearchRepoHasWiki :: !(Maybe Bool)

codeSearchRepoHasPages :: !(Maybe Bool)

codeSearchRepoHasDownloads :: !(Maybe Bool)

codeSearchRepoArchived :: !Bool

codeSearchRepoDisabled :: !Bool

codeSearchRepoPushedAt :: !(Maybe UTCTime)

this is Nothing for new repositories

codeSearchRepoCreatedAt :: !(Maybe UTCTime)

codeSearchRepoUpdatedAt :: !(Maybe UTCTime)

codeSearchRepoPermissions :: !(Maybe RepoPermissions)

Repository permissions as they relate to the authenticated user.

data RepoPermissions

Repository permissions, as they relate to the authenticated user.

Returned by for example currentUserReposR

Constructors

RepoPermissions

Fields

repoPermissionAdmin :: !Bool

repoPermissionPush :: !Bool

repoPermissionPull :: !Bool

data RepoRef

Constructors

RepoRef

Fields

repoRefOwner :: !SimpleOwner

repoRefRepo :: !(Name Repo)

data NewRepo

Constructors

NewRepo

Fields

newRepoName :: !(Name Repo)

newRepoDescription :: !(Maybe Text)

newRepoHomepage :: !(Maybe Text)

newRepoPrivate :: !(Maybe Bool)

newRepoHasIssues :: !(Maybe Bool)

newRepoHasProjects :: !(Maybe Bool)

newRepoHasWiki :: !(Maybe Bool)

newRepoAutoInit :: !(Maybe Bool)

newRepoGitignoreTemplate :: !(Maybe Text)

newRepoLicenseTemplate :: !(Maybe Text)

newRepoAllowSquashMerge :: !(Maybe Bool)

newRepoAllowMergeCommit :: !(Maybe Bool)

newRepoAllowRebaseMerge :: !(Maybe Bool)

newRepo :: Name Repo -> NewRepo

data EditRepo

Constructors

EditRepo

Fields

editName :: !(Maybe (Name Repo))

editDescription :: !(Maybe Text)

editHomepage :: !(Maybe Text)

editPrivate :: !(Maybe Bool)

editHasIssues :: !(Maybe Bool)

editHasProjects :: !(Maybe Bool)

editHasWiki :: !(Maybe Bool)

editDefaultBranch :: !(Maybe Text)

editAllowSquashMerge :: !(Maybe Bool)

editAllowMergeCommit :: !(Maybe Bool)

editAllowRebaseMerge :: !(Maybe Bool)

editArchived :: !(Maybe Bool)

data RepoPublicity

Filter the list of the user's repos using any of these constructors.

Constructors

RepoPublicityAll
All repos accessible to the user.

RepoPublicityOwner
Only repos owned by the user.

RepoPublicityPublic
Only public repos.

RepoPublicityPrivate
Only private repos.

RepoPublicityMember
Only repos to which the user is a member but not an owner.

type Languages = HashMap Language Int

The value is the number of bytes of code written in that language.

newtype Language

A programming language.

Constructors

Language Text

getLanguage :: Language -> Text

data Contributor

Constructors

KnownContributor !Int !URL !(Name User) !URL !(Id User) !Text
An existing Github user, with their number of contributions, avatar URL, login, URL, ID, and Gravatar ID.

AnonymousContributor !Int !Text
An unknown Github user with their number of contributions and recorded name.

contributorToSimpleUser :: Contributor -> Maybe SimpleUser

data CollaboratorPermission

The permission of a collaborator on a repository. See https://developer.github.com/v3/repos/collaborators/#review-a-users-permission-level

Constructors

CollaboratorPermissionAdmin

CollaboratorPermissionWrite

CollaboratorPermissionRead

CollaboratorPermissionNone

data CollaboratorWithPermission

A collaborator and its permission on a repository. See https://developer.github.com/v3/repos/collaborators/#review-a-users-permission-level

Constructors

CollaboratorWithPermission SimpleUser CollaboratorPermission

data ArchiveFormat

Constructors

ArchiveFormatTarball
".tar.gz" format

ArchiveFormatZipball
".zip" format