GitHub.Data.Request

Request

type Request = GenRequest ('MtJSON :: MediaType Type)

Most requests ask for JSON.

data GenRequest (mt :: MediaType Type) (rw :: RW) a where

Github request data type.

rw describes whether authentication is required. It's required for non-GET requests.

mt describes the media type, i.e. how the response should be interpreted.

a is the result type

Note: Request is not Functor on purpose.

Constructors

Query :: forall (mt :: MediaType Type) (rw :: RW) a. Paths -> QueryString -> GenRequest mt rw a

PagedQuery :: forall a (t :: Type -> Type) b (mt :: MediaType Type) (rw :: RW). (a ~ t b, Foldable t, Semigroup a) => Paths -> QueryString -> FetchCount -> GenRequest mt rw a

Command
Command

Fields

:: forall (mt :: MediaType Type) a. CommandMethod

command

-> Paths

path

-> ByteString

body

-> GenRequest mt 'RW a

Smart constructors

query :: forall (mt :: RW) a. Paths -> QueryString -> Request mt a

pagedQuery :: forall a (mt :: RW). FromJSON a => Paths -> QueryString -> FetchCount -> Request mt (Vector a)

command :: CommandMethod -> Paths -> ByteString -> Request 'RW a

Auxiliary types

data RW

Type used as with DataKinds to tag whether requests need authentication or aren't read-only.

Constructors

RO
Read-only, doesn't necessarily requires authentication

RA
Read authenticated

RW
Read-write, requires authentication

data CommandMethod

Http method of requests with body.

Constructors

Post

Patch

Put

Delete

toMethod :: CommandMethod -> Method

data FetchCount

PagedQuery returns just some results, using this data we can specify how many pages we want to fetch.

Constructors

FetchAtLeast !Word

FetchAll

FetchPage PageParams

data PageParams

Params for specifying the precise page and items per page.

Constructors

PageParams

Fields

pageParamsPerPage :: Maybe Int

pageParamsPage :: Maybe Int

data PageLinks

PagedQuery returns just some results, using this data we can specify how many pages we want to fetch.

Constructors

PageLinks

Fields

pageLinksPrev :: Maybe URI

pageLinksNext :: Maybe URI

pageLinksLast :: Maybe URI

pageLinksFirst :: Maybe URI

data MediaType a

Constructors

MtJSON
application/vnd.github.v3+json

MtRaw
application/vnd.github.v3.raw https://developer.github.com/v3/media/#raw-1

MtDiff
application/vnd.github.v3.diff https://developer.github.com/v3/media/#diff

MtPatch
application/vnd.github.v3.patch https://developer.github.com/v3/media/#patch

MtSha
application/vnd.github.v3.sha https://developer.github.com/v3/media/#sha

MtStar
application/vnd.github.v3.star+json https://developer.github.com/v3/activity/starring/#alternative-response-with-star-creation-timestamps-1

MtRedirect
https://developer.github.com/v3/repos/contents/#get-archive-link

MtStatus
Parse status

MtUnit
Always succeeds

MtPreview a
Some other (preview) type; this is an extension point.

type Paths = [Text]

class IsPathPart a where

Methods

toPathPart :: a -> Text

type QueryString = [(ByteString, Maybe ByteString)]

Request query string

type Count = Int

Count of elements