GitHub.Endpoints.Activity.Starring

Description

The repo starring API as described on https://developer.github.com/v3/activity/starring/.

Documentation

stargazersFor :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector SimpleUser))

The list of users that have starred the specified Github repo.

userInfoFor' Nothing "mike-burns"

stargazersForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector SimpleUser)

List Stargazers. See https://developer.github.com/v3/activity/starring/#list-stargazers

reposStarredBy :: Maybe Auth -> Name Owner -> IO (Either Error (Vector Repo))

All the public repos starred by the specified user.

reposStarredBy Nothing "croaky"

reposStarredByR :: Name Owner -> FetchCount -> Request k (Vector Repo)

List repositories being starred. See https://developer.github.com/v3/activity/starring/#list-repositories-being-starred

myStarred :: Auth -> IO (Either Error (Vector Repo))

All the repos starred by the authenticated user.

myStarredR :: FetchCount -> Request RA (Vector Repo)

All the repos starred by the authenticated user. See https://developer.github.com/v3/activity/starring/#list-repositories-being-starred

myStarredAcceptStar :: Auth -> IO (Either Error (Vector RepoStarred))

All the repos starred by the authenticated user.

myStarredAcceptStarR :: FetchCount -> Request RA (Vector RepoStarred)

All the repos starred by the authenticated user. See https://developer.github.com/v3/activity/starring/#alternative-response-with-star-creation-timestamps-1

starRepo :: Auth -> Name Owner -> Name Repo -> IO (Either Error ())

Star a repo by the authenticated user.

starRepoR :: Name Owner -> Name Repo -> Request RW ()

Star a repo by the authenticated user. See https://developer.github.com/v3/activity/starring/#star-a-repository

unstarRepo :: Auth -> Name Owner -> Name Repo -> IO (Either Error ())

Unstar a repo by the authenticated user.

unstarRepoR :: Name Owner -> Name Repo -> Request RW ()

Unstar a repo by the authenticated user. See https://developer.github.com/v3/activity/starring/#unstar-a-repository

module GitHub.Data