GitHub.Endpoints.Activity.Watching

Description

The repo watching API as described on https://developer.github.com/v3/activity/watching/.

Documentation

watchersFor :: Name Owner -> Name Repo -> IO (Either Error (Vector SimpleUser))

The list of users that are watching the specified Github repo.

watchersFor "thoughtbot" "paperclip"

watchersFor' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector SimpleUser))

The list of users that are watching the specified Github repo. With authentication

watchersFor' (Just (User (user, password))) "thoughtbot" "paperclip"

watchersForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector SimpleUser)

List watchers. See https://developer.github.com/v3/activity/watching/#list-watchers

reposWatchedBy :: Name Owner -> IO (Either Error (Vector Repo))

All the public repos watched by the specified user.

reposWatchedBy "croaky"

reposWatchedBy' :: Maybe Auth -> Name Owner -> IO (Either Error (Vector Repo))

All the public repos watched by the specified user. With authentication

reposWatchedBy' (Just (User (user, password))) "croaky"

reposWatchedByR :: Name Owner -> FetchCount -> Request k (Vector Repo)

List repositories being watched. See https://developer.github.com/v3/activity/watching/#list-repositories-being-watched

module GitHub.Data