GitHub.Endpoints.Gists

Description

The gists API as described at http://developer.github.com/v3/gists/.

Documentation

gists :: Name Owner -> IO (Either Error (Vector Gist))

The list of all public gists created by the user.

gists "mike-burns"

gists' :: Maybe Auth -> Name Owner -> IO (Either Error (Vector Gist))

The list of all gists created by the user

gists' (Just ("github-username", "github-password")) "mike-burns"

gistsR :: Name Owner -> FetchCount -> Request k (Vector Gist)

List gists. See https://developer.github.com/v3/gists/#list-gists

gist :: Name Gist -> IO (Either Error Gist)

A specific gist, given its id.

gist "225074"

gist' :: Maybe Auth -> Name Gist -> IO (Either Error Gist)

A specific gist, given its id, with authentication credentials

gist' (Just ("github-username", "github-password")) "225074"

gistR :: Name Gist -> Request k Gist

Query a single gist. See https://developer.github.com/v3/gists/#get-a-single-gist

starGist :: Auth -> Name Gist -> IO (Either Error ())

Star a gist by the authenticated user.

starGist ("github-username", "github-password") "225074"

starGistR :: Name Gist -> Request RW ()

Star a gist by the authenticated user. See https://developer.github.com/v3/gists/#star-a-gist

unstarGist :: Auth -> Name Gist -> IO (Either Error ())

Unstar a gist by the authenticated user.

unstarGist ("github-username", "github-password") "225074"

unstarGistR :: Name Gist -> Request RW ()

Unstar a gist by the authenticated user. See https://developer.github.com/v3/gists/#unstar-a-gist

deleteGist :: Auth -> Name Gist -> IO (Either Error ())

Delete a gist by the authenticated user.

deleteGist ("github-username", "github-password") "225074"

deleteGistR :: Name Gist -> Request RW ()

Delete a gist by the authenticated user. See https://developer.github.com/v3/gists/#delete-a-gist

module GitHub.Data