GitHub.Endpoints.Gists.Comments

Description

The loving comments people have left on Gists, described on http://developer.github.com/v3/gists/comments/.

Documentation

commentsOn :: Name Gist -> IO (Either Error (Vector GistComment))

All the comments on a Gist, given the Gist ID.

commentsOn "1174060"

commentsOnR :: Name Gist -> FetchCount -> Request k (Vector GistComment)

List comments on a gist. See https://developer.github.com/v3/gists/comments/#list-comments-on-a-gist

comment :: Id GistComment -> IO (Either Error GistComment)

A specific comment, by the comment ID.

comment (Id 62449)

gistCommentR :: Id GistComment -> Request k GistComment

Query a single comment. See https://developer.github.com/v3/gists/comments/#get-a-single-comment

module GitHub.Data