GitHub.Endpoints.GitData.References

Description

The underlying git references on a Github repo, exposed for the world to see. The git internals documentation will also prove handy for understanding these. API documentation at http://developer.github.com/v3/git/refs/.

Documentation

reference :: Name Owner -> Name Repo -> Name GitReference -> IO (Either Error GitReference)

A single reference by the ref name.

reference "mike-burns" "github" "heads/master"

reference' :: Maybe Auth -> Name Owner -> Name Repo -> Name GitReference -> IO (Either Error GitReference)

A single reference by the ref name.

reference' (Just ("github-username", "github-password")) "mike-burns" "github" "heads/master"

referenceR :: Name Owner -> Name Repo -> Name GitReference -> Request k GitReference

Query a reference. See https://developer.github.com/v3/git/refs/#get-a-reference

references :: Name Owner -> Name Repo -> IO (Either Error (Vector GitReference))

The history of references for a repo.

references "mike-burns" "github"

references' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector GitReference))

The history of references for a repo.

references "mike-burns" "github"

referencesR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector GitReference)

Query all References. See https://developer.github.com/v3/git/refs/#get-all-references

createReference :: Auth -> Name Owner -> Name Repo -> NewGitReference -> IO (Either Error GitReference)

Create a reference.

createReferenceR :: Name Owner -> Name Repo -> NewGitReference -> Request RW GitReference

Create a reference. See https://developer.github.com/v3/git/refs/#create-a-reference

namespacedReferences :: Name Owner -> Name Repo -> Text -> IO (Either Error [GitReference])

Limited references by a namespace.

namespacedReferences "thoughtbot" "paperclip" "tags"

module GitHub.Data