GitHub.Endpoints.Issues

Description

The issues API as described on http://developer.github.com/v3/issues/.

Documentation

currentUserIssuesR :: IssueMod -> FetchCount -> Request RA (Vector Issue)

See https://developer.github.com/v3/issues/#list-issues.

organizationIssuesR :: Name Organization -> IssueMod -> FetchCount -> Request k (Vector Issue)

See https://developer.github.com/v3/issues/#list-issues.

issue :: Name Owner -> Name Repo -> Id Issue -> IO (Either Error Issue)

Details on a specific issue, given the repo owner and name, and the issue number.

issue "thoughtbot" "paperclip" (Id "462")

issue' :: Maybe Auth -> Name Owner -> Name Repo -> Id Issue -> IO (Either Error Issue)

Details on a specific issue, given the repo owner and name, and the issue number.'

issue' (Just ("github-username", "github-password")) "thoughtbot" "paperclip" "462"

issueR :: Name Owner -> Name Repo -> Id Issue -> Request k Issue

Query a single issue. See https://developer.github.com/v3/issues/#get-a-single-issue

issuesForRepo :: Name Owner -> Name Repo -> IssueRepoMod -> IO (Either Error (Vector Issue))

All issues for a repo (given the repo owner and name), with optional restrictions as described in the IssueRepoMod data type.

issuesForRepo "thoughtbot" "paperclip" [NoMilestone, OnlyClosed, Mentions "jyurek", Ascending]

issuesForRepo' :: Maybe Auth -> Name Owner -> Name Repo -> IssueRepoMod -> IO (Either Error (Vector Issue))

All issues for a repo (given the repo owner and name), with optional restrictions as described in the IssueRepoMod data type.

issuesForRepo' (Just ("github-username", "github-password")) "thoughtbot" "paperclip" [NoMilestone, OnlyClosed, Mentions "jyurek", Ascending]

issuesForRepoR :: Name Owner -> Name Repo -> IssueRepoMod -> FetchCount -> Request k (Vector Issue)

List issues for a repository. See https://developer.github.com/v3/issues/#list-issues-for-a-repository

createIssue :: Auth -> Name Owner -> Name Repo -> NewIssue -> IO (Either Error Issue)

Create a new issue.

createIssue (User (user, password)) user repo
 (newIssue "some_repo") {...}

createIssueR :: Name Owner -> Name Repo -> NewIssue -> Request RW Issue

Create an issue. See https://developer.github.com/v3/issues/#create-an-issue

newIssue :: Text -> NewIssue

editIssue :: Auth -> Name Owner -> Name Repo -> Id Issue -> EditIssue -> IO (Either Error Issue)

Edit an issue.

editIssue (User (user, password)) user repo issue
 editOfIssue {...}

editIssueR :: Name Owner -> Name Repo -> Id Issue -> EditIssue -> Request RW Issue

Edit an issue. See https://developer.github.com/v3/issues/#edit-an-issue

editOfIssue :: EditIssue

module GitHub.Data