GitHub.Endpoints.Issues.Labels

Description

The API for dealing with labels on Github issues as described on http://developer.github.com/v3/issues/labels/.

Documentation

labelsOnRepoR :: forall (k :: RW). Name Owner -> Name Repo -> FetchCount -> Request k (Vector IssueLabel)

List all labels for this repository. See https://developer.github.com/v3/issues/labels/#list-all-labels-for-this-repository

labelR :: forall (k :: RW). Name Owner -> Name Repo -> Name IssueLabel -> Request k IssueLabel

Query a single label. See https://developer.github.com/v3/issues/labels/#get-a-single-label

createLabelR :: Name Owner -> Name Repo -> NewIssueLabel -> Request 'RW IssueLabel

Create a label. See https://developer.github.com/v3/issues/labels/#create-a-label

updateLabelR

Arguments

:: Name Owner

-> Name Repo

-> Name IssueLabel
old label name

-> UpdateIssueLabel
new label

-> Request 'RW IssueLabel

Update a label. See https://developer.github.com/v3/issues/labels/#update-a-label

deleteLabelR :: Name Owner -> Name Repo -> Name IssueLabel -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Delete a label. See https://developer.github.com/v3/issues/labels/#delete-a-label

labelsOnIssueR :: forall (k :: RW). Name Owner -> Name Repo -> Id Issue -> FetchCount -> Request k (Vector IssueLabel)

List labels on an issue. See https://developer.github.com/v3/issues/labels/#list-labels-on-an-issue

addLabelsToIssueR :: Foldable f => Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> Request 'RW (Vector IssueLabel)

Add lables to an issue. See https://developer.github.com/v3/issues/labels/#add-labels-to-an-issue

removeLabelFromIssueR :: Name Owner -> Name Repo -> Id Issue -> Name IssueLabel -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Remove a label from an issue. See https://developer.github.com/v3/issues/labels/#remove-a-label-from-an-issue

replaceAllLabelsForIssueR :: Foldable f => Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> Request 'RW (Vector IssueLabel)

Replace all labels on an issue. See https://developer.github.com/v3/issues/labels/#replace-all-labels-for-an-issue

Sending an empty list will remove all labels from the issue.

removeAllLabelsFromIssueR :: Name Owner -> Name Repo -> Id Issue -> GenRequest ('MtUnit :: MediaType Type) 'RW ()

Remove all labels from an issue. See https://developer.github.com/v3/issues/labels/#remove-all-labels-from-an-issue

labelsOnMilestoneR :: forall (k :: RW). Name Owner -> Name Repo -> Id Milestone -> FetchCount -> Request k (Vector IssueLabel)

Query labels for every issue in a milestone. See https://developer.github.com/v3/issues/labels/#get-labels-for-every-issue-in-a-milestone

module GitHub.Data