GitHub.Endpoints.Organizations

Description

The orgs API as described on http://developer.github.com/v3/orgs/.

Documentation

publicOrganizationsFor :: Name User -> IO (Either Error (Vector SimpleOrganization))

List user organizations. The public organizations for a user, given the user's login.

publicOrganizationsFor "mike-burns"

publicOrganizationsFor' :: Maybe Auth -> Name User -> IO (Either Error (Vector SimpleOrganization))

The public organizations for a user, given the user's login, with authorization

publicOrganizationsFor' (Just ("github-username", "github-password")) "mike-burns"

publicOrganizationsForR :: Name User -> FetchCount -> Request k (Vector SimpleOrganization)

List user organizations. See https://developer.github.com/v3/orgs/#list-user-organizations

publicOrganization :: Name Organization -> IO (Either Error Organization)

Query an organization. Details on a public organization. Takes the organization's login.

publicOrganization "thoughtbot"

publicOrganization' :: Maybe Auth -> Name Organization -> IO (Either Error Organization)

Details on a public organization. Takes the organization's login.

publicOrganization' (Just ("github-username", "github-password")) "thoughtbot"

publicOrganizationR :: Name Organization -> Request k Organization

Query an organization. See https://developer.github.com/v3/orgs/#get-an-organization

module GitHub.Data