GitHub.Endpoints.Organizations.Members

Description

The organization members API as described on http://developer.github.com/v3/orgs/members/.

Documentation

membersOfR :: forall (k :: RW). Name Organization -> FetchCount -> Request k (Vector SimpleUser)

All the users who are members of the specified organization.

See https://developer.github.com/v3/orgs/members/#members-list

membersOfWithR :: forall (k :: RW). Name Organization -> OrgMemberFilter -> OrgMemberRole -> FetchCount -> Request k (Vector SimpleUser)

membersOfR with filters.

See https://developer.github.com/v3/orgs/members/#members-list

isMemberOfR :: forall (rw :: RW). Name User -> Name Organization -> GenRequest ('MtStatus :: MediaType Type) rw Bool

Check if a user is a member of an organization.

See https://developer.github.com/v3/orgs/members/#check-membership

orgInvitationsR :: Name Organization -> FetchCount -> Request 'RA (Vector Invitation)

List pending organization invitations

See https://developer.github.com/v3/orgs/members/#list-pending-organization-invitations

orgMembershipR :: Name User -> Name Organization -> Request 'RA Membership

Get user membership information in an organization

See https://docs.github.com/en/rest/orgs/members#get-organization-membership-for-a-user

module GitHub.Data