GitHub.Endpoints.Organizations.Teams

Description

The Owner teams API as described on http://developer.github.com/v3/orgs/teams/.

Documentation

teamsOf :: Name Organization -> IO (Either Error (Vector SimpleTeam))

List the public teams of an Owner.

teamsOf "thoughtbot"

teamsOf' :: Maybe Auth -> Name Organization -> IO (Either Error (Vector SimpleTeam))

List teams. List the teams of an Owner. When authenticated, lists private teams visible to the authenticated user. When unauthenticated, lists only public teams for an Owner.

teamsOf' (Just $ OAuth "token") "thoughtbot"

teamsOfR :: Name Organization -> FetchCount -> Request k (Vector SimpleTeam)

List teams. See https://developer.github.com/v3/orgs/teams/#list-teams

teamInfoFor :: Id Team -> IO (Either Error Team)

The information for a single team, by team id.

teamInfoFor' (Just $ OAuth "token") 1010101

teamInfoFor' :: Maybe Auth -> Id Team -> IO (Either Error Team)

The information for a single team, by team id. With authentication

teamInfoFor' (Just $ OAuth "token") 1010101

teamInfoForR :: Id Team -> Request k Team

Query team. See https://developer.github.com/v3/orgs/teams/#get-team

createTeamFor' :: Auth -> Name Organization -> CreateTeam -> IO (Either Error Team)

Create a team under an Owner

createTeamFor' (OAuth "token") "Owner" (CreateTeam "newteamname" "some description" [] PermssionPull)

createTeamForR :: Name Organization -> CreateTeam -> Request RW Team

Create team. See https://developer.github.com/v3/orgs/teams/#create-team

editTeam' :: Auth -> Id Team -> EditTeam -> IO (Either Error Team)

Edit a team, by id.

editTeamFor'

editTeamR :: Id Team -> EditTeam -> Request RW Team

Edit team. See https://developer.github.com/v3/orgs/teams/#edit-team

deleteTeam' :: Auth -> Id Team -> IO (Either Error ())

Delete a team, by id.

deleteTeam' (OAuth "token") 1010101

deleteTeamR :: Id Team -> Request RW ()

Delete team.

See https://developer.github.com/v3/orgs/teams/#delete-team

listTeamMembersR :: Id Team -> TeamMemberRole -> FetchCount -> Request RA (Vector SimpleUser)

List team members.

See https://developer.github.com/v3/orgs/teams/#list-team-members

listTeamRepos :: Id Team -> IO (Either Error (Vector Repo))

Retrieve repositories for a team.

listTeamRepos (GitHub.mkTeamId team_id)

listTeamRepos' :: Maybe Auth -> Id Team -> IO (Either Error (Vector Repo))

The repositories of a single team, by team id. With authentication

listTeamRepos' (Just $ GitHub.OAuth token) (GitHub.mkTeamId team_id)

listTeamReposR :: Id Team -> FetchCount -> Request k (Vector Repo)

Query team repositories. See https://developer.github.com/v3/orgs/teams/#list-team-repos

addOrUpdateTeamRepo' :: Auth -> Id Team -> Name Organization -> Name Repo -> Permission -> IO (Either Error ())

Add a repository to a team or update the permission on the repository.

addOrUpdateTeamRepo' (OAuth "token") 1010101 "mburns" (Just PermissionPull)

addOrUpdateTeamRepoR :: Id Team -> Name Organization -> Name Repo -> Permission -> Request RW ()

Add or update a team repository. See https://developer.github.com/v3/orgs/teams/#add-or-update-team-repository

teamMembershipInfoFor :: Id Team -> Name Owner -> IO (Either Error TeamMembership)

Retrieve team mebership information for a user.

teamMembershipInfoFor 1010101 "mburns"

teamMembershipInfoFor' :: Maybe Auth -> Id Team -> Name Owner -> IO (Either Error TeamMembership)

Retrieve team mebership information for a user. With authentication

teamMembershipInfoFor' (Just $ OAuth "token") 1010101 "mburns"

teamMembershipInfoForR :: Id Team -> Name Owner -> Request k TeamMembership

Query team membership. See <https://developer.github.com/v3/orgs/teams/#get-team-membership

addTeamMembershipFor' :: Auth -> Id Team -> Name Owner -> Role -> IO (Either Error TeamMembership)

Add (or invite) a member to a team.

addTeamMembershipFor' (OAuth "token") 1010101 "mburns" RoleMember

addTeamMembershipForR :: Id Team -> Name Owner -> Role -> Request RW TeamMembership

Add team membership. See https://developer.github.com/v3/orgs/teams/#add-team-membership

deleteTeamMembershipFor' :: Auth -> Id Team -> Name Owner -> IO (Either Error ())

Delete a member of a team.

deleteTeamMembershipFor' (OAuth "token") 1010101 "mburns"

deleteTeamMembershipForR :: Id Team -> Name Owner -> Request RW ()

Remove team membership. See https://developer.github.com/v3/orgs/teams/#remove-team-membership

listTeamsCurrent' :: Auth -> IO (Either Error (Vector Team))

List teams for current authenticated user

listTeamsCurrent' (OAuth "token")

listTeamsCurrentR :: FetchCount -> Request RA (Vector Team)

List user teams. See https://developer.github.com/v3/orgs/teams/#list-user-teams

module GitHub.Data