GitHub.Endpoints.PullRequests

Description

The pull requests API as documented at http://developer.github.com/v3/pulls/.

Documentation

pullRequestsForR :: forall (k :: RW). Name Owner -> Name Repo -> PullRequestMod -> FetchCount -> Request k (Vector SimplePullRequest)

List pull requests. See https://developer.github.com/v3/pulls/#list-pull-requests

pullRequestR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> Request k PullRequest

Query a single pull request. See https://developer.github.com/v3/pulls/#get-a-single-pull-request

pullRequestDiffR :: forall (rw :: RW). Name Owner -> Name Repo -> IssueNumber -> GenRequest ('MtDiff :: MediaType Type) rw ByteString

Query a single pull request to obtain the diff See https://developer.github.com/v3/pulls/#get-a-single-pull-request

pullRequestPatchR :: forall (rw :: RW). Name Owner -> Name Repo -> IssueNumber -> GenRequest ('MtPatch :: MediaType Type) rw ByteString

Query a single pull request to obtain the patch See https://developer.github.com/v3/pulls/#get-a-single-pull-request

createPullRequestR :: Name Owner -> Name Repo -> CreatePullRequest -> Request 'RW PullRequest

Create a pull request. See https://developer.github.com/v3/pulls/#create-a-pull-request

updatePullRequestR :: Name Owner -> Name Repo -> IssueNumber -> EditPullRequest -> Request 'RW PullRequest

Update a pull request. See https://developer.github.com/v3/pulls/#update-a-pull-request

pullRequestCommitsR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> FetchCount -> Request k (Vector Commit)

List commits on a pull request. See https://developer.github.com/v3/pulls/#list-commits-on-a-pull-request

pullRequestFilesR :: forall (k :: RW). Name Owner -> Name Repo -> IssueNumber -> FetchCount -> Request k (Vector File)

List pull requests files. See https://developer.github.com/v3/pulls/#list-pull-requests-files

isPullRequestMergedR :: forall (rw :: RW). Name Owner -> Name Repo -> IssueNumber -> GenRequest ('MtStatus :: MediaType Type) rw Bool

Query if a pull request has been merged. See https://developer.github.com/v3/pulls/#get-if-a-pull-request-has-been-merged

mergePullRequestR :: Name Owner -> Name Repo -> IssueNumber -> Maybe Text -> GenRequest ('MtStatus :: MediaType Type) 'RW MergeResult

Merge a pull request (Merge Button). https://developer.github.com/v3/pulls/#merge-a-pull-request-merge-button

module GitHub.Data