GitHub.Endpoints.PullRequests.Comments

Description

The pull request review comments API as described at http://developer.github.com/v3/pulls/comments/.

Documentation

pullRequestCommentsIO :: Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error (Vector Comment))

All the comments on a pull request with the given ID.

pullRequestComments "thoughtbot" "factory_girl" (Id 256)

pullRequestCommentsR :: Name Owner -> Name Repo -> Id PullRequest -> FetchCount -> Request k (Vector Comment)

List comments on a pull request. See https://developer.github.com/v3/pulls/comments/#list-comments-on-a-pull-request

pullRequestComment :: Name Owner -> Name Repo -> Id Comment -> IO (Either Error Comment)

One comment on a pull request, by the comment's ID.

pullRequestComment "thoughtbot" "factory_girl" (Id 301819)

pullRequestCommentR :: Name Owner -> Name Repo -> Id Comment -> Request k Comment

Query a single comment. See https://developer.github.com/v3/pulls/comments/#get-a-single-comment

module GitHub.Data