GitHub.Endpoints.PullRequests.Reviews

Description

The reviews API as described on http://developer.github.com/v3/pulls/reviews/.

Documentation

pullRequestReviewsR :: Name Owner -> Name Repo -> Id PullRequest -> FetchCount -> Request k (Vector Review)

List reviews for a pull request. See https://developer.github.com/v3/pulls/reviews/#list-reviews-on-a-pull-request

pullRequestReviews :: Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error (Vector Review))

All reviews for a pull request given the repo owner, repo name and the pull request id.

pullRequestReviews "thoughtbot" "paperclip" (Id 101)

pullRequestReviews' :: Maybe Auth -> Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error (Vector Review))

All reviews for a pull request given the repo owner, repo name and the pull request id. With authentication.

pullRequestReviews' (Just ("github-username", "github-password")) "thoughtbot" "paperclip" (Id 101)

pullRequestReviewR :: Name Owner -> Name Repo -> Id PullRequest -> Id Review -> Request k Review

Query a single pull request review. see https://developer.github.com/v3/pulls/reviews/#get-a-single-review

pullRequestReview :: Name Owner -> Name Repo -> Id PullRequest -> Id Review -> IO (Either Error Review)

A detailed review on a pull request given the repo owner, repo name, pull request id and review id.

pullRequestReview "thoughtbot" "factory_girl" (Id 301819) (Id 332)

pullRequestReview' :: Maybe Auth -> Name Owner -> Name Repo -> Id PullRequest -> Id Review -> IO (Either Error Review)

A detailed review on a pull request given the repo owner, repo name, pull request id and review id. With authentication.

pullRequestReview' (Just ("github-username", "github-password"))

"thoughtbot" "factory_girl" (Id 301819) (Id 332)

pullRequestReviewCommentsR :: Name Owner -> Name Repo -> Id PullRequest -> Id Review -> Request k [ReviewComment]

Query the comments for a single pull request review. see https://developer.github.com/v3/pulls/reviews/#get-comments-for-a-single-review

pullRequestReviewCommentsIO :: Name Owner -> Name Repo -> Id PullRequest -> Id Review -> IO (Either Error [ReviewComment])

All comments for a review on a pull request given the repo owner, repo name, pull request id and review id.

pullRequestReviewComments "thoughtbot" "factory_girl" (Id 301819) (Id 332)

pullRequestReviewCommentsIO' :: Maybe Auth -> Name Owner -> Name Repo -> Id PullRequest -> Id Review -> IO (Either Error [ReviewComment])

All comments for a review on a pull request given the repo owner, repo name, pull request id and review id. With authentication.

pullRequestReviewComments' (Just ("github-username", "github-password")) "thoughtbot" "factory_girl" (Id 301819) (Id 332)

module GitHub.Data