GitHub.Endpoints.Repos

Description

The Github Repos API, as documented at http://developer.github.com/v3/repos/

Querying repositories

currentUserRepos :: Auth -> RepoPublicity -> IO (Either Error (Vector Repo))

List your repositories.

currentUserReposR :: RepoPublicity -> FetchCount -> Request k (Vector Repo)

List your repositories. See https://developer.github.com/v3/repos/#list-your-repositories

userRepos :: Name Owner -> RepoPublicity -> IO (Either Error (Vector Repo))

The repos for a user, by their login. Can be restricted to just repos they own, are a member of, or publicize. Private repos will return empty list.

userRepos "mike-burns" All

userRepos' :: Maybe Auth -> Name Owner -> RepoPublicity -> IO (Either Error (Vector Repo))

The repos for a user, by their login. With authentication.

userRepos' (Just (BasicAuth (user, password))) "mike-burns" All

userReposR :: Name Owner -> RepoPublicity -> FetchCount -> Request k (Vector Repo)

List user repositories. See https://developer.github.com/v3/repos/#list-user-repositories

organizationRepos :: Name Organization -> IO (Either Error (Vector Repo))

The repos for an organization, by the organization name.

organizationRepos "thoughtbot"

organizationRepos' :: Maybe Auth -> Name Organization -> RepoPublicity -> IO (Either Error (Vector Repo))

The repos for an organization, by the organization name. With authentication.

organizationRepos (Just (BasicAuth (user, password))) "thoughtbot" All

organizationReposR :: Name Organization -> RepoPublicity -> FetchCount -> Request k (Vector Repo)

List organization repositories. See https://developer.github.com/v3/repos/#list-organization-repositories

repository :: Name Owner -> Name Repo -> IO (Either Error Repo)

Details on a specific repo, given the owner and repo name.

repository "mike-burns" "github"

repository' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error Repo)

Details on a specific repo, given the owner and repo name. With authentication.

repository' (Just (BasicAuth (user, password))) "mike-burns" "github"

repositoryR :: Name Owner -> Name Repo -> Request k Repo

Query single repository. See https://developer.github.com/v3/repos/#get

contributors :: Name Owner -> Name Repo -> IO (Either Error (Vector Contributor))

The contributors to a repo, given the owner and repo name.

contributors "thoughtbot" "paperclip"

contributors' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector Contributor))

The contributors to a repo, given the owner and repo name. With authentication.

contributors' (Just (BasicAuth (user, password))) "thoughtbot" "paperclip"

contributorsR

Arguments

:: Name Owner

-> Name Repo

-> Bool
Include anonymous

-> FetchCount

-> Request k (Vector Contributor)

List contributors. See https://developer.github.com/v3/repos/#list-contributors

contributorsWithAnonymous :: Name Owner -> Name Repo -> IO (Either Error (Vector Contributor))

The contributors to a repo, including anonymous contributors (such as deleted users or git commits with unknown email addresses), given the owner and repo name.

contributorsWithAnonymous "thoughtbot" "paperclip"

contributorsWithAnonymous' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector Contributor))

The contributors to a repo, including anonymous contributors (such as deleted users or git commits with unknown email addresses), given the owner and repo name. With authentication.

contributorsWithAnonymous' (Just (BasicAuth (user, password))) "thoughtbot" "paperclip"

languagesFor :: Name Owner -> Name Repo -> IO (Either Error Languages)

The programming languages used in a repo along with the number of characters written in that language. Takes the repo owner and name.

languagesFor "mike-burns" "ohlaunch"

languagesFor' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error Languages)

The programming languages used in a repo along with the number of characters written in that language. Takes the repo owner and name. With authentication.

languagesFor' (Just (BasicAuth (user, password))) "mike-burns" "ohlaunch"

languagesForR :: Name Owner -> Name Repo -> Request k Languages

List languages. See https://developer.github.com/v3/repos/#list-languages

tagsFor :: Name Owner -> Name Repo -> IO (Either Error (Vector Tag))

The git tags on a repo, given the repo owner and name.

tagsFor "thoughtbot" "paperclip"

tagsFor' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector Tag))

The git tags on a repo, given the repo owner and name. With authentication.

tagsFor' (Just (BasicAuth (user, password))) "thoughtbot" "paperclip"

tagsForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector Tag)

List tags. See https://developer.github.com/v3/repos/#list-tags

branchesFor :: Name Owner -> Name Repo -> IO (Either Error (Vector Branch))

The git branches on a repo, given the repo owner and name.

branchesFor "thoughtbot" "paperclip"

branchesFor' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector Branch))

The git branches on a repo, given the repo owner and name. With authentication.

branchesFor' (Just (BasicAuth (user, password))) "thoughtbot" "paperclip"

branchesForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector Branch)

List branches. See https://developer.github.com/v3/repos/#list-branches

Create

createRepo' :: Auth -> NewRepo -> IO (Either Error Repo)

Create a new repository.

createRepo' (BasicAuth (user, password)) (newRepo "some_repo") {newRepoHasIssues = Just False}

createRepoR :: NewRepo -> Request RW Repo

Create a new repository. See https://developer.github.com/v3/repos/#create

createOrganizationRepo' :: Auth -> Name Organization -> NewRepo -> IO (Either Error Repo)

Create a new repository for an organization.

createOrganizationRepo (BasicAuth (user, password)) "thoughtbot" (newRepo "some_repo") {newRepoHasIssues = Just False}

createOrganizationRepoR :: Name Organization -> NewRepo -> Request RW Repo

Create a new repository for an organization. See https://developer.github.com/v3/repos/#create

forkExistingRepoR :: Name Owner -> Name Repo -> Maybe (Name Owner) -> Request RW Repo

Fork an existing repository. See https://developer.github.com/v3/repos/forks/#create-a-fork TODO: The third paramater (an optional Organisation) is not used yet.

Edit

editRepo

Arguments

:: Auth

-> Name Owner
owner

-> Name Repo
repository name

-> EditRepo

-> IO (Either Error Repo)

Edit an existing repository.

editRepo (BasicAuth (user, password)) "some_user" "some_repo" def {editDescription = Just "some description"}

editRepoR :: Name Owner -> Name Repo -> EditRepo -> Request RW Repo

Edit an existing repository. See https://developer.github.com/v3/repos/#edit

Delete

deleteRepo :: Auth -> Name Owner -> Name Repo -> IO (Either Error ())

Delete an existing repository.

deleteRepo (BasicAuth (user, password)) "thoughtbot" "some_repo"

deleteRepoR :: Name Owner -> Name Repo -> Request RW ()

Data

module GitHub.Data