GitHub.Endpoints.Repos.Collaborators

Description

The repo collaborators API as described on http://developer.github.com/v3/repos/collaborators/.

Documentation

collaboratorsOn :: Name Owner -> Name Repo -> IO (Either Error (Vector SimpleUser))

All the users who have collaborated on a repo.

collaboratorsOn "thoughtbot" "paperclip"

collaboratorsOn' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector SimpleUser))

All the users who have collaborated on a repo. With authentication.

collaboratorsOnR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector SimpleUser)

List collaborators. See https://developer.github.com/v3/repos/collaborators/#list-collaborators

isCollaboratorOn

Arguments

:: Maybe Auth

-> Name Owner
Repository owner

-> Name Repo
Repository name

-> Name User
Collaborator?

-> IO (Either Error Bool)

Whether the user is collaborating on a repo. Takes the user in question, the user who owns the repo, and the repo name.

isCollaboratorOn Nothing "mike-burns" "thoughtbot" "paperclip"
isCollaboratorOn Nothing "johnson" "thoughtbot" "paperclip"

isCollaboratorOnR

Arguments

:: Name Owner
Repository owner

-> Name Repo
Repository name

-> Name User
Collaborator?

-> Request k Bool

Check if a user is a collaborator. See https://developer.github.com/v3/repos/collaborators/#check-if-a-user-is-a-collaborator

module GitHub.Data