GitHub.Endpoints.Repos.Contents

Description

The Github Repo Contents API, as documented at https://developer.github.com/v3/repos/contents/

Querying contents

contentsFor :: Name Owner -> Name Repo -> Text -> Maybe Text -> IO (Either Error Content)

The contents of a file or directory in a repo, given the repo owner, name, and path to the file

contentsFor "thoughtbot" "paperclip" "README.md"

contentsFor' :: Maybe Auth -> Name Owner -> Name Repo -> Text -> Maybe Text -> IO (Either Error Content)

The contents of a file or directory in a repo, given the repo owner, name, and path to the file With Authentication

contentsFor' (Just (BasicAuth (user, password))) "thoughtbot" "paperclip" "README.md" Nothing

contentsForR

Arguments

:: Name Owner

-> Name Repo

-> Text
file or directory

-> Maybe Text
Git commit

-> Request k Content

readmeFor :: Name Owner -> Name Repo -> IO (Either Error Content)

The contents of a README file in a repo, given the repo owner and name

readmeFor "thoughtbot" "paperclip"

readmeFor' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error Content)

The contents of a README file in a repo, given the repo owner and name With Authentication

readmeFor' (Just (BasicAuth (user, password))) "thoughtbot" "paperclip"

readmeForR :: Name Owner -> Name Repo -> Request k Content

archiveFor :: Name Owner -> Name Repo -> ArchiveFormat -> Maybe Text -> IO (Either Error URI)

The archive of a repo, given the repo owner, name, and archive type

archiveFor "thoughtbot" "paperclip" ArchiveFormatTarball Nothing

archiveFor' :: Maybe Auth -> Name Owner -> Name Repo -> ArchiveFormat -> Maybe Text -> IO (Either Error URI)

The archive of a repo, given the repo owner, name, and archive type With Authentication

archiveFor' (Just (BasicAuth (user, password))) "thoughtbot" "paperclip" ArchiveFormatTarball Nothing

archiveForR

Arguments

:: Name Owner

-> Name Repo

-> ArchiveFormat
The type of archive to retrieve

-> Maybe Text
Git commit

-> Request k URI

Create

createFile

Arguments

:: Auth

-> Name Owner
owner

-> Name Repo
repository name

-> CreateFile

-> IO (Either Error ContentResult)

Create a file.

createFileR :: Name Owner -> Name Repo -> CreateFile -> Request RW ContentResult

Create a file. See https://developer.github.com/v3/repos/contents/#create-a-file

Update

updateFile

Arguments

:: Auth

-> Name Owner
owner

-> Name Repo
repository name

-> UpdateFile

-> IO (Either Error ContentResult)

Update a file.

updateFileR :: Name Owner -> Name Repo -> UpdateFile -> Request RW ContentResult

Update a file. See https://developer.github.com/v3/repos/contents/#update-a-file

Delete

deleteFile

Arguments

:: Auth

-> Name Owner
owner

-> Name Repo
repository name

-> DeleteFile

-> IO (Either Error ())

Delete a file.

deleteFileR :: Name Owner -> Name Repo -> DeleteFile -> Request RW ()

Delete a file. See https://developer.github.com/v3/repos/contents/#delete-a-file

module GitHub.Data