GitHub.Endpoints.Users

Description

The Github Users API, as described at http://developer.github.com/v3/users/.

Documentation

userInfoFor :: Name User -> IO (Either Error User)

The information for a single user, by login name.

userInfoFor "mike-burns"

userInfoFor' :: Maybe Auth -> Name User -> IO (Either Error User)

The information for a single user, by login name. With authentification

userInfoFor' (Just ("github-username", "github-password")) "mike-burns"

userInfoForR :: Name User -> Request k User

Query a single user. See https://developer.github.com/v3/users/#get-a-single-user

ownerInfoForR :: Name Owner -> Request k Owner

Query a single user or an organization. See https://developer.github.com/v3/users/#get-a-single-user

userInfoCurrent' :: Auth -> IO (Either Error User)

Retrieve information about the user associated with the supplied authentication.

userInfoCurrent' (OAuth "...")

userInfoCurrentR :: Request RA User

Query the authenticated user. See https://developer.github.com/v3/users/#get-the-authenticated-user

module GitHub.Data