GitHub.Request

Description

This module provides data types and helper methods, which makes possible to build alternative API request intepreters in addition to provided IO functions.

Simple example using operational package. See samples/Operational/Operational.hs

type GithubMonad a = Program (GH.Request 'False) a

-- | Intepret GithubMonad value into IO
runMonad :: Manager -> GH.Auth -> GithubMonad a -> ExceptT GH.Error IO a
runMonad mgr auth m = case view m of
   Return a   -> return a
   req :>>= k -> do
       b <- ExceptT $ GH.executeRequestWithMgr mgr auth req
       runMonad mgr auth (k b)

-- | Lift request into Monad
githubRequest :: GH.Request 'False a -> GithubMonad a
githubRequest = singleton

A convenient execution of requests

github :: (AuthMethod am, GitHubRW req res) => am -> req -> res

A convenience function to turn functions returning Request rw x, into functions returning IO (Either Error x).

>>> :t \auth -> github auth userInfoForR
\auth -> github auth userInfoForR
  :: AuthMethod am => am -> Name User -> IO (Either Error User)

>>> :t github pullRequestsForR
\auth -> github auth pullRequestsForR
  :: AuthMethod am =>
     am
     -> Name Owner
     -> Name Repo
     -> PullRequestMod
     -> FetchCount
     -> IO (Either Error (Data.Vector.Vector SimplePullRequest))

github' :: GitHubRO req res => req -> res

Like github' but for RO i.e. read-only requests. Note that GitHub has low request limit for non-authenticated requests.

>>> :t github' userInfoForR
github' userInfoForR :: Name User -> IO (Either Error User)

class GitHubRW req res | req -> res

A type-class implementing github.

Minimal complete definition

githubImpl

class GitHubRO req res | req -> res

A type-class implementing github'.

Minimal complete definition

githubImpl'

Types

type Request = GenRequest ('MtJSON :: MediaType Type)

Most requests ask for JSON.

data GenRequest (mt :: MediaType Type) (rw :: RW) a where

Github request data type.

rw describes whether authentication is required. It's required for non-GET requests.

mt describes the media type, i.e. how the response should be interpreted.

a is the result type

Note: Request is not Functor on purpose.

Constructors

Query :: forall (mt :: MediaType Type) (rw :: RW) a. Paths -> QueryString -> GenRequest mt rw a

PagedQuery :: forall a (t :: Type -> Type) b (mt :: MediaType Type) (rw :: RW). (a ~ t b, Foldable t, Semigroup a) => Paths -> QueryString -> FetchCount -> GenRequest mt rw a

Command
Command

Fields

:: forall (mt :: MediaType Type) a. CommandMethod

command

-> Paths

path

-> ByteString

body

-> GenRequest mt 'RW a

data CommandMethod

Http method of requests with body.

Constructors

Post

Patch

Put

Delete

toMethod :: CommandMethod -> Method

type Paths = [Text]

type QueryString = [(ByteString, Maybe ByteString)]

Request query string

Request execution in IO

executeRequest :: forall am (mt :: MediaType Type) a (rw :: RW). (AuthMethod am, ParseResponse mt a) => am -> GenRequest mt rw a -> IO (Either Error a)

Execute Request in IO

executeRequestWithMgr :: forall am (mt :: MediaType Type) a (rw :: RW). (AuthMethod am, ParseResponse mt a) => Manager -> am -> GenRequest mt rw a -> IO (Either Error a)

Like executeRequest but with provided Manager.

executeRequestWithMgrAndRes :: forall am (mt :: MediaType Type) a (rw :: RW). (AuthMethod am, ParseResponse mt a) => Manager -> am -> GenRequest mt rw a -> IO (Either Error (Response a))

Execute request and return the last received Response.

Since: 0.24

executeRequest' :: forall (mt :: MediaType Type) a. ParseResponse mt a => GenRequest mt 'RO a -> IO (Either Error a)

Like executeRequest but without authentication.

executeRequestWithMgr' :: forall (mt :: MediaType Type) a. ParseResponse mt a => Manager -> GenRequest mt 'RO a -> IO (Either Error a)

Like executeRequestWithMgr but without authentication.

executeRequestMaybe :: forall am (mt :: MediaType Type) a. (AuthMethod am, ParseResponse mt a) => Maybe am -> GenRequest mt 'RO a -> IO (Either Error a)

Helper for picking between executeRequest and executeRequest'.

The use is discouraged.

unsafeDropAuthRequirements :: forall (mt :: MediaType Type) (rw' :: RW) a (rw :: RW). GenRequest mt rw' a -> GenRequest mt rw a

Partial function to drop authentication need.

Helpers

class Accept (mt :: MediaType Type) where

Minimal complete definition

Nothing

Methods

contentType :: Tagged mt ByteString

modifyRequest :: Tagged mt (Request -> Request)

class Accept mt => ParseResponse (mt :: MediaType Type) a where

Methods

parseResponse :: MonadError Error m => Request -> Response ByteString -> Tagged mt (m a)

makeHttpRequest :: forall am (mt :: MediaType Type) (rw :: RW) a m. (AuthMethod am, MonadThrow m, Accept mt) => Maybe am -> GenRequest mt rw a -> m Request

Create http-client Request.

for PagedQuery, the initial request is created.

for Status, the Request for underlying Request is created, status checking is modifying accordingly.

parseStatus :: MonadError Error m => StatusMap a -> Status -> m a

Helper for handling of RequestStatus.

parseStatus :: StatusMap a -> Status -> Either Error a

parsePageLinks :: Response a -> PageLinks

Parse the PageLinks from an HTTP response, where the information is encoded in the Link header.

type StatusMap a = [(Int, a)]

getNextUrl :: Response a -> Maybe URI

Query Link header with rel=next from the request headers.

performPagedRequest

Arguments

:: forall a m (mt :: MediaType Type). (ParseResponse mt a, Semigroup a, MonadCatch m, MonadError Error m)

=> (Request -> m (Response ByteString))
httpLbs analogue

-> (a -> Bool)
predicate to continue iteration

-> Request
initial request

-> Tagged mt (m (Response a))

Helper for making paginated requests. Responses, a are combined monoidally.

The result is wrapped in the last received Response.

performPagedRequest :: (FromJSON a, Semigroup a)
                    => (Request -> ExceptT Error IO (Response ByteString))
                    -> (a -> Bool)
                    -> Request
                    -> ExceptT Error IO (Response a)

parseResponseJSON :: (FromJSON a, MonadError Error m) => Response ByteString -> m a

Parse API response.

parseResponse :: FromJSON a => Response ByteString -> Either Error a

Preview

class PreviewAccept p where

Minimal complete definition

previewContentType

Methods

previewContentType :: Tagged ('MtPreview p) ByteString

previewModifyRequest :: Tagged ('MtPreview p) (Request -> Request)

class PreviewAccept p => PreviewParseResponse p a where

Methods

previewParseResponse :: MonadError Error m => Request -> Response ByteString -> Tagged ('MtPreview p) (m a)

SSL

This always exist, independently of openssl configuration flag. They change accordingly, to make use of the library simpler.

withOpenSSL :: IO a -> IO a

tlsManagerSettings :: ManagerSettings

Default TLS-enabled manager settings