GitHub

Description

This module re-exports all request constructrors and data definitions from this package.

See GitHub.Request module for executing Request, or other modules of this package (e.g. GitHub.Users) for already composed versions.

The missing endpoints lists show which endpoints we know are missing, there might be more.

Activity

See https://developer.github.com/v3/activity/

Events

See https://developer.github.com/v3/activity/events/#events

repositoryEventsR :: Name Owner -> Name Repo -> FetchCount -> Request RO (Vector Event)

List repository events. See https://developer.github.com/v3/activity/events/#list-repository-events

userEventsR :: Name User -> FetchCount -> Request RO (Vector Event)

List user public events. See https://developer.github.com/v3/activity/events/#list-public-events-performed-by-a-user

Starring

See https://developer.github.com/v3/activity/starring/

Missing endpoints:

Check if you are starring a repository

stargazersForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector SimpleUser)

List Stargazers. See https://developer.github.com/v3/activity/starring/#list-stargazers

reposStarredByR :: Name Owner -> FetchCount -> Request k (Vector Repo)

List repositories being starred. See https://developer.github.com/v3/activity/starring/#list-repositories-being-starred

myStarredR :: FetchCount -> Request RA (Vector Repo)

All the repos starred by the authenticated user. See https://developer.github.com/v3/activity/starring/#list-repositories-being-starred

myStarredAcceptStarR :: FetchCount -> Request RA (Vector RepoStarred)

All the repos starred by the authenticated user. See https://developer.github.com/v3/activity/starring/#alternative-response-with-star-creation-timestamps-1

starRepoR :: Name Owner -> Name Repo -> Request RW ()

Star a repo by the authenticated user. See https://developer.github.com/v3/activity/starring/#star-a-repository

unstarRepoR :: Name Owner -> Name Repo -> Request RW ()

Unstar a repo by the authenticated user. See https://developer.github.com/v3/activity/starring/#unstar-a-repository

Watching

See https://developer.github.com/v3/activity/

Missing endpoints:

Query a Repository Subscription

Set a Repository Subscription

Delete a Repository Subscription

watchersForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector SimpleUser)

List watchers. See https://developer.github.com/v3/activity/watching/#list-watchers

reposWatchedByR :: Name Owner -> FetchCount -> Request k (Vector Repo)

List repositories being watched. See https://developer.github.com/v3/activity/watching/#list-repositories-being-watched

Gists

See https://developer.github.com/v3/gists/

Missing endpoints:

Query a specific revision of a gist

Create a gist

Edit a gist

List gist commits

Check if a gist is starred

Fork a gist

List gist forks

gistsR :: Name Owner -> FetchCount -> Request k (Vector Gist)

List gists. See https://developer.github.com/v3/gists/#list-gists

gistR :: Name Gist -> Request k Gist

Query a single gist. See https://developer.github.com/v3/gists/#get-a-single-gist

starGistR :: Name Gist -> Request RW ()

Star a gist by the authenticated user. See https://developer.github.com/v3/gists/#star-a-gist

unstarGistR :: Name Gist -> Request RW ()

Unstar a gist by the authenticated user. See https://developer.github.com/v3/gists/#unstar-a-gist

deleteGistR :: Name Gist -> Request RW ()

Delete a gist by the authenticated user. See https://developer.github.com/v3/gists/#delete-a-gist

Comments

See https://developer.github.com/v3/gists/comments/

Missing endpoints: * Create a comment * Edit a comment * Delete a comment

commentsOnR :: Name Gist -> FetchCount -> Request k (Vector GistComment)

List comments on a gist. See https://developer.github.com/v3/gists/comments/#list-comments-on-a-gist

gistCommentR :: Id GistComment -> Request k GistComment

Query a single comment. See https://developer.github.com/v3/gists/comments/#get-a-single-comment

Git Data

See https://developer.github.com/v3/git/

Blobs

See https://developer.github.com/v3/git/blobs/

blobR :: Name Owner -> Name Repo -> Name Blob -> Request k Blob

Query a blob. See https://developer.github.com/v3/git/blobs/#get-a-blob

Commits

See https://developer.github.com/v3/git/commits/

gitCommitR :: Name Owner -> Name Repo -> Name GitCommit -> Request k GitCommit

Query a commit. See https://developer.github.com/v3/git/commits/#get-a-commit

References

See https://developer.github.com/v3/git/refs/

referenceR :: Name Owner -> Name Repo -> Name GitReference -> Request k GitReference

Query a reference. See https://developer.github.com/v3/git/refs/#get-a-reference

referencesR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector GitReference)

Query all References. See https://developer.github.com/v3/git/refs/#get-all-references

createReferenceR :: Name Owner -> Name Repo -> NewGitReference -> Request RW GitReference

Create a reference. See https://developer.github.com/v3/git/refs/#create-a-reference

Trees

See https://developer.github.com/v3/git/trees/

treeR :: Name Owner -> Name Repo -> Name Tree -> Request k Tree

Query a Tree. See https://developer.github.com/v3/git/trees/#get-a-tree

nestedTreeR :: Name Owner -> Name Repo -> Name Tree -> Request k Tree

Query a Tree Recursively. See https://developer.github.com/v3/git/trees/#get-a-tree-recursively

Issues

See https://developer.github.com/v3/issues/

currentUserIssuesR :: IssueMod -> FetchCount -> Request RA (Vector Issue)

See https://developer.github.com/v3/issues/#list-issues.

organizationIssuesR :: Name Organization -> IssueMod -> FetchCount -> Request k (Vector Issue)

See https://developer.github.com/v3/issues/#list-issues.

issueR :: Name Owner -> Name Repo -> Id Issue -> Request k Issue

Query a single issue. See https://developer.github.com/v3/issues/#get-a-single-issue

issuesForRepoR :: Name Owner -> Name Repo -> IssueRepoMod -> FetchCount -> Request k (Vector Issue)

List issues for a repository. See https://developer.github.com/v3/issues/#list-issues-for-a-repository

createIssueR :: Name Owner -> Name Repo -> NewIssue -> Request RW Issue

Create an issue. See https://developer.github.com/v3/issues/#create-an-issue

editIssueR :: Name Owner -> Name Repo -> Id Issue -> EditIssue -> Request RW Issue

Edit an issue. See https://developer.github.com/v3/issues/#edit-an-issue

Comments

See https://developer.github.com/v3/issues/comments/

commentR :: Name Owner -> Name Repo -> Id Comment -> Request k IssueComment

Query a single comment. See https://developer.github.com/v3/issues/comments/#get-a-single-comment

commentsR :: Name Owner -> Name Repo -> Id Issue -> FetchCount -> Request k (Vector IssueComment)

List comments on an issue. See https://developer.github.com/v3/issues/comments/#list-comments-on-an-issue

createCommentR :: Name Owner -> Name Repo -> Id Issue -> Text -> Request RW Comment

Create a comment. See https://developer.github.com/v3/issues/comments/#create-a-comment

deleteCommentR :: Name Owner -> Name Repo -> Id Comment -> Request RW ()

Delete a comment. See https://developer.github.com/v3/issues/comments/#delete-a-comment

editCommentR :: Name Owner -> Name Repo -> Id Comment -> Text -> Request RW Comment

Edit a comment. See https://developer.github.com/v3/issues/comments/#edit-a-comment

Events

See https://developer.github.com/v3/issues/events/

eventsForIssueR :: Name Owner -> Name Repo -> Id Issue -> FetchCount -> Request k (Vector IssueEvent)

List events for an issue. See https://developer.github.com/v3/issues/events/#list-events-for-an-issue

eventsForRepoR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector IssueEvent)

List events for a repository. See https://developer.github.com/v3/issues/events/#list-events-for-a-repository

eventR :: Name Owner -> Name Repo -> Id IssueEvent -> Request k IssueEvent

Query a single event. See https://developer.github.com/v3/issues/events/#get-a-single-event

Labels

See https://developer.github.com/v3/issues/labels/

labelsOnRepoR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector IssueLabel)

List all labels for this repository. See https://developer.github.com/v3/issues/labels/#list-all-labels-for-this-repository

labelR :: Name Owner -> Name Repo -> Name IssueLabel -> Request k IssueLabel

Query a single label. See https://developer.github.com/v3/issues/labels/#get-a-single-label

createLabelR :: Name Owner -> Name Repo -> Name IssueLabel -> String -> Request RW IssueLabel

Create a label. See https://developer.github.com/v3/issues/labels/#create-a-label

updateLabelR

Arguments

:: Name Owner

-> Name Repo

-> Name IssueLabel
old label name

-> Name IssueLabel
new label name

-> String
new color

-> Request RW IssueLabel

Update a label. See https://developer.github.com/v3/issues/labels/#update-a-label

deleteLabelR :: Name Owner -> Name Repo -> Name IssueLabel -> Request RW ()

Delete a label. See https://developer.github.com/v3/issues/labels/#delete-a-label

labelsOnIssueR :: Name Owner -> Name Repo -> Id Issue -> FetchCount -> Request k (Vector IssueLabel)

List labels on an issue. See https://developer.github.com/v3/issues/labels/#list-labels-on-an-issue

addLabelsToIssueR :: Foldable f => Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> Request RW (Vector IssueLabel)

Add lables to an issue. See https://developer.github.com/v3/issues/labels/#add-labels-to-an-issue

removeLabelFromIssueR :: Name Owner -> Name Repo -> Id Issue -> Name IssueLabel -> Request RW ()

Remove a label from an issue. See https://developer.github.com/v3/issues/labels/#remove-a-label-from-an-issue

replaceAllLabelsForIssueR :: Foldable f => Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> Request RW (Vector IssueLabel)

Replace all labels on an issue. See https://developer.github.com/v3/issues/labels/#replace-all-labels-for-an-issue

Sending an empty list will remove all labels from the issue.

removeAllLabelsFromIssueR :: Name Owner -> Name Repo -> Id Issue -> Request RW ()

Remove all labels from an issue. See https://developer.github.com/v3/issues/labels/#remove-all-labels-from-an-issue

labelsOnMilestoneR :: Name Owner -> Name Repo -> Id Milestone -> FetchCount -> Request k (Vector IssueLabel)

Query labels for every issue in a milestone. See https://developer.github.com/v3/issues/labels/#get-labels-for-every-issue-in-a-milestone

Milestone

See https://developer.github.com/v3/issues/milestones/

Missing endpoints:

Create a milestone

Update a milestone

Delete a milestone

milestonesR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector Milestone)

List milestones for a repository. See https://developer.github.com/v3/issues/milestones/#list-milestones-for-a-repository

milestoneR :: Name Owner -> Name Repo -> Id Milestone -> Request k Milestone

Query a single milestone. See https://developer.github.com/v3/issues/milestones/#get-a-single-milestone

Organizations

See https://developer.github.com/v3/orgs/

Missing endpoints:

List your organizations

List all organizations

Edit an organization

publicOrganizationsForR :: Name User -> FetchCount -> Request k (Vector SimpleOrganization)

List user organizations. See https://developer.github.com/v3/orgs/#list-user-organizations

publicOrganizationR :: Name Organization -> Request k Organization

Query an organization. See https://developer.github.com/v3/orgs/#get-an-organization

Members

See https://developer.github.com/v3/orgs/members/

Missing endpoints: All except Members List and Check Membership

membersOfR :: Name Organization -> FetchCount -> Request k (Vector SimpleUser)

All the users who are members of the specified organization.

See https://developer.github.com/v3/orgs/members/#members-list

membersOfWithR :: Name Organization -> OrgMemberFilter -> OrgMemberRole -> FetchCount -> Request k (Vector SimpleUser)

membersOfR with filters.

See https://developer.github.com/v3/orgs/members/#members-list

isMemberOfR :: Name User -> Name Organization -> Request k Bool

Check if a user is a member of an organization.

See https://developer.github.com/v3/orgs/members/#check-membership

orgInvitationsR :: Name Organization -> FetchCount -> Request RA (Vector Invitation)

List pending organization invitations

See https://developer.github.com/v3/orgs/members/#list-pending-organization-invitations

Teams

See https://developer.github.com/v3/orgs/teams/

Missing endpoints:

Query team member (deprecated)

Add team member (deprecated)

Remove team member (deprecated)

Check if a team manages a repository

Add team repository

Remove team repository

teamsOfR :: Name Organization -> FetchCount -> Request k (Vector SimpleTeam)

List teams. See https://developer.github.com/v3/orgs/teams/#list-teams

teamInfoForR :: Id Team -> Request k Team

Query team. See https://developer.github.com/v3/orgs/teams/#get-team

createTeamForR :: Name Organization -> CreateTeam -> Request RW Team

Create team. See https://developer.github.com/v3/orgs/teams/#create-team

editTeamR :: Id Team -> EditTeam -> Request RW Team

Edit team. See https://developer.github.com/v3/orgs/teams/#edit-team

deleteTeamR :: Id Team -> Request RW ()

Delete team.

See https://developer.github.com/v3/orgs/teams/#delete-team

listTeamMembersR :: Id Team -> TeamMemberRole -> FetchCount -> Request RA (Vector SimpleUser)

List team members.

See https://developer.github.com/v3/orgs/teams/#list-team-members

listTeamReposR :: Id Team -> FetchCount -> Request k (Vector Repo)

Query team repositories. See https://developer.github.com/v3/orgs/teams/#list-team-repos

teamMembershipInfoForR :: Id Team -> Name Owner -> Request k TeamMembership

Query team membership. See <https://developer.github.com/v3/orgs/teams/#get-team-membership

addTeamMembershipForR :: Id Team -> Name Owner -> Role -> Request RW TeamMembership

Add team membership. See https://developer.github.com/v3/orgs/teams/#add-team-membership

deleteTeamMembershipForR :: Id Team -> Name Owner -> Request RW ()

Remove team membership. See https://developer.github.com/v3/orgs/teams/#remove-team-membership

listTeamsCurrentR :: FetchCount -> Request RA (Vector Team)

List user teams. See https://developer.github.com/v3/orgs/teams/#list-user-teams

Pull Requests

See https://developer.github.com/v3/pulls/

pullRequestsForR :: Name Owner -> Name Repo -> PullRequestMod -> FetchCount -> Request k (Vector SimplePullRequest)

List pull requests. See https://developer.github.com/v3/pulls/#list-pull-requests

pullRequestR :: Name Owner -> Name Repo -> Id PullRequest -> Request k PullRequest

Query a single pull request. See https://developer.github.com/v3/pulls/#get-a-single-pull-request

createPullRequestR :: Name Owner -> Name Repo -> CreatePullRequest -> Request RW PullRequest

Create a pull request. See https://developer.github.com/v3/pulls/#create-a-pull-request

updatePullRequestR :: Name Owner -> Name Repo -> Id PullRequest -> EditPullRequest -> Request RW PullRequest

Update a pull request. See https://developer.github.com/v3/pulls/#update-a-pull-request

pullRequestCommitsR :: Name Owner -> Name Repo -> Id PullRequest -> FetchCount -> Request k (Vector Commit)

List commits on a pull request. See https://developer.github.com/v3/pulls/#list-commits-on-a-pull-request

pullRequestFilesR :: Name Owner -> Name Repo -> Id PullRequest -> FetchCount -> Request k (Vector File)

List pull requests files. See https://developer.github.com/v3/pulls/#list-pull-requests-files

isPullRequestMergedR :: Name Owner -> Name Repo -> Id PullRequest -> Request k Bool

Query if a pull request has been merged. See https://developer.github.com/v3/pulls/#get-if-a-pull-request-has-been-merged

mergePullRequestR :: Name Owner -> Name Repo -> Id PullRequest -> Maybe Text -> Request RW MergeResult

Merge a pull request (Merge Button). https://developer.github.com/v3/pulls/#merge-a-pull-request-merge-button

Review comments

See https://developer.github.com/v3/pulls/comments/

Missing endpoints:

List comments in a repository

Create a comment

Edit a comment

Delete a comment

pullRequestCommentsR :: Name Owner -> Name Repo -> Id PullRequest -> FetchCount -> Request k (Vector Comment)

List comments on a pull request. See https://developer.github.com/v3/pulls/comments/#list-comments-on-a-pull-request

pullRequestCommentR :: Name Owner -> Name Repo -> Id Comment -> Request k Comment

Query a single comment. See https://developer.github.com/v3/pulls/comments/#get-a-single-comment

Pull request reviews

See https://developer.github.com/v3/pulls/reviews/

Missing endpoints:

Delete a pending review

Create a pull request review

Submit a pull request review

Dismiss a pull request review

pullRequestReviewsR :: Name Owner -> Name Repo -> Id PullRequest -> FetchCount -> Request k (Vector Review)

List reviews for a pull request. See https://developer.github.com/v3/pulls/reviews/#list-reviews-on-a-pull-request

pullRequestReviews :: Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error (Vector Review))

All reviews for a pull request given the repo owner, repo name and the pull request id.

pullRequestReviews "thoughtbot" "paperclip" (Id 101)

pullRequestReviews' :: Maybe Auth -> Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error (Vector Review))

All reviews for a pull request given the repo owner, repo name and the pull request id. With authentication.

pullRequestReviews' (Just ("github-username", "github-password")) "thoughtbot" "paperclip" (Id 101)

pullRequestReviewR :: Name Owner -> Name Repo -> Id PullRequest -> Id Review -> Request k Review

Query a single pull request review. see https://developer.github.com/v3/pulls/reviews/#get-a-single-review

pullRequestReview :: Name Owner -> Name Repo -> Id PullRequest -> Id Review -> IO (Either Error Review)

A detailed review on a pull request given the repo owner, repo name, pull request id and review id.

pullRequestReview "thoughtbot" "factory_girl" (Id 301819) (Id 332)

pullRequestReview' :: Maybe Auth -> Name Owner -> Name Repo -> Id PullRequest -> Id Review -> IO (Either Error Review)

A detailed review on a pull request given the repo owner, repo name, pull request id and review id. With authentication.

pullRequestReview' (Just ("github-username", "github-password"))

"thoughtbot" "factory_girl" (Id 301819) (Id 332)

pullRequestReviewCommentsR :: Name Owner -> Name Repo -> Id PullRequest -> Id Review -> Request k [ReviewComment]

Query the comments for a single pull request review. see https://developer.github.com/v3/pulls/reviews/#get-comments-for-a-single-review

pullRequestReviewCommentsIO :: Name Owner -> Name Repo -> Id PullRequest -> Id Review -> IO (Either Error [ReviewComment])

All comments for a review on a pull request given the repo owner, repo name, pull request id and review id.

pullRequestReviewComments "thoughtbot" "factory_girl" (Id 301819) (Id 332)

pullRequestReviewCommentsIO' :: Maybe Auth -> Name Owner -> Name Repo -> Id PullRequest -> Id Review -> IO (Either Error [ReviewComment])

All comments for a review on a pull request given the repo owner, repo name, pull request id and review id. With authentication.

pullRequestReviewComments' (Just ("github-username", "github-password")) "thoughtbot" "factory_girl" (Id 301819) (Id 332)

Repositories

See https://developer.github.com/v3/repos/

Missing endpoints:

List all public repositories

List Teams

Query Branch

Enabling and disabling branch protection

currentUserReposR :: RepoPublicity -> FetchCount -> Request k (Vector Repo)

List your repositories. See https://developer.github.com/v3/repos/#list-your-repositories

userReposR :: Name Owner -> RepoPublicity -> FetchCount -> Request k (Vector Repo)

List user repositories. See https://developer.github.com/v3/repos/#list-user-repositories

organizationReposR :: Name Organization -> RepoPublicity -> FetchCount -> Request k (Vector Repo)

List organization repositories. See https://developer.github.com/v3/repos/#list-organization-repositories

repositoryR :: Name Owner -> Name Repo -> Request k Repo

Query single repository. See https://developer.github.com/v3/repos/#get

contributorsR

Arguments

:: Name Owner

-> Name Repo

-> Bool
Include anonymous

-> FetchCount

-> Request k (Vector Contributor)

List contributors. See https://developer.github.com/v3/repos/#list-contributors

languagesForR :: Name Owner -> Name Repo -> Request k Languages

List languages. See https://developer.github.com/v3/repos/#list-languages

tagsForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector Tag)

List tags. See https://developer.github.com/v3/repos/#list-tags

branchesForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector Branch)

List branches. See https://developer.github.com/v3/repos/#list-branches

Collaborators

See https://developer.github.com/v3/repos/collaborators/

collaboratorsOnR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector SimpleUser)

List collaborators. See https://developer.github.com/v3/repos/collaborators/#list-collaborators

isCollaboratorOnR

Arguments

:: Name Owner
Repository owner

-> Name Repo
Repository name

-> Name User
Collaborator?

-> Request k Bool

Check if a user is a collaborator. See https://developer.github.com/v3/repos/collaborators/#check-if-a-user-is-a-collaborator

Comments

See https://developer.github.com/v3/repos/comments/

Missing endpoints:

Create a commit comment

Update a commit comment

Delete a commit comment

commentsForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector Comment)

List commit comments for a repository. See https://developer.github.com/v3/repos/comments/#list-commit-comments-for-a-repository

commitCommentsForR :: Name Owner -> Name Repo -> Name Commit -> FetchCount -> Request k (Vector Comment)

List comments for a single commit. See https://developer.github.com/v3/repos/comments/#list-comments-for-a-single-commit

commitCommentForR :: Name Owner -> Name Repo -> Id Comment -> Request k Comment

Query a single commit comment. See https://developer.github.com/v3/repos/comments/#get-a-single-commit-comment

Commits

See https://developer.github.com/v3/repos/commits/

commitsForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector Commit)

List commits on a repository. See https://developer.github.com/v3/repos/commits/#list-commits-on-a-repository

commitsWithOptionsForR :: Name Owner -> Name Repo -> FetchCount -> [CommitQueryOption] -> Request k (Vector Commit)

List commits on a repository. See https://developer.github.com/v3/repos/commits/#list-commits-on-a-repository

commitR :: Name Owner -> Name Repo -> Name Commit -> Request k Commit

Query a single commit. See https://developer.github.com/v3/repos/commits/#get-a-single-commit

diffR :: Name Owner -> Name Repo -> Name Commit -> Name Commit -> Request k Diff

Compare two commits. See https://developer.github.com/v3/repos/commits/#compare-two-commits

Forks

See https://developer.github.com/v3/repos/forks/

Missing endpoints:

Create a fork

forksForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector Repo)

List forks. See https://developer.github.com/v3/repos/forks/#list-forks

Webhooks

See https://developer.github.com/v3/repos/hooks/

webhooksForR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector RepoWebhook)

List hooks. See https://developer.github.com/v3/repos/hooks/#list-hooks

webhookForR :: Name Owner -> Name Repo -> Id RepoWebhook -> Request k RepoWebhook

Query single hook. See https://developer.github.com/v3/repos/hooks/#get-single-hook

createRepoWebhookR :: Name Owner -> Name Repo -> NewRepoWebhook -> Request RW RepoWebhook

Create a hook. See https://developer.github.com/v3/repos/hooks/#create-a-hook

editRepoWebhookR :: Name Owner -> Name Repo -> Id RepoWebhook -> EditRepoWebhook -> Request RW RepoWebhook

Edit a hook. See https://developer.github.com/v3/repos/hooks/#edit-a-hook

testPushRepoWebhookR :: Name Owner -> Name Repo -> Id RepoWebhook -> Request RW Bool

Test a push hook. See https://developer.github.com/v3/repos/hooks/#test-a-push-hook

pingRepoWebhookR :: Name Owner -> Name Repo -> Id RepoWebhook -> Request RW Bool

Ping a hook. See https://developer.github.com/v3/repos/hooks/#ping-a-hook

deleteRepoWebhookR :: Name Owner -> Name Repo -> Id RepoWebhook -> Request RW ()

Delete a hook. See https://developer.github.com/v3/repos/hooks/#delete-a-hook

Releases

releasesR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector Release)

List releases for a repository. See https://developer.github.com/v3/repos/releases/#list-releases-for-a-repository

releaseR :: Name Owner -> Name Repo -> Id Release -> Request k Release

Get a single release. See https://developer.github.com/v3/repos/releases/#get-a-single-release

latestReleaseR :: Name Owner -> Name Repo -> Request k Release

Get the latest release. See https://developer.github.com/v3/repos/releases/#get-the-latest-release

releaseByTagNameR :: Name Owner -> Name Repo -> Text -> Request k Release

Get a release by tag name See https://developer.github.com/v3/repos/releases/#get-a-release-by-tag-name

Search

See https://developer.github.com/v3/search/

Missing endpoints:

Search users

searchReposR :: Text -> Request k (SearchResult Repo)

Search repositories. See https://developer.github.com/v3/search/#search-repositories

searchCodeR :: Text -> Request k (SearchResult Code)

Search code. See https://developer.github.com/v3/search/#search-code

searchIssuesR :: Text -> Request k (SearchResult Issue)

Search issues. See https://developer.github.com/v3/search/#search-issues

Users

See https://developer.github.com/v3/users/

Missing endpoints:

Update the authenticated user

Query all users

userInfoForR :: Name User -> Request k User

Query a single user. See https://developer.github.com/v3/users/#get-a-single-user

ownerInfoForR :: Name Owner -> Request k Owner

Query a single user or an organization. See https://developer.github.com/v3/users/#get-a-single-user

userInfoCurrentR :: Request RA User

Query the authenticated user. See https://developer.github.com/v3/users/#get-the-authenticated-user

Emails

See https://developer.github.com/v3/users/emails/

Missing endpoints:

Add email address(es)

Delete email address(es)

Toggle primary email visibility

currentUserEmailsR :: FetchCount -> Request RA (Vector Email)

List email addresses. See https://developer.github.com/v3/users/emails/#list-email-addresses-for-a-user

currentUserPublicEmailsR :: FetchCount -> Request RA (Vector Email)

List public email addresses. See https://developer.github.com/v3/users/emails/#list-public-email-addresses-for-a-user

Followers

See https://developer.github.com/v3/users/followers/

Missing endpoints:

Check if you are following a user

Check if one user follows another

Follow a user

Unfollow a user

usersFollowingR :: Name User -> FetchCount -> Request k (Vector SimpleUser)

List followers of a user. See https://developer.github.com/v3/users/followers/#list-followers-of-a-user

usersFollowedByR :: Name User -> FetchCount -> Request k (Vector SimpleUser)

List users followed by another user. See https://developer.github.com/v3/users/followers/#list-users-followed-by-another-user

Statuses

See https://developer.github.com/v3/repos/statuses/

createStatusR :: Name Owner -> Name Repo -> Name Commit -> NewStatus -> Request RW Status

Create a new status See https://developer.github.com/v3/repos/statuses/#create-a-status

statusesForR :: Name Owner -> Name Repo -> Name Commit -> FetchCount -> Request RW (Vector Status)

All statuses for a commit See https://developer.github.com/v3/repos/statuses/#list-statuses-for-a-specific-ref

statusForR :: Name Owner -> Name Repo -> Name Commit -> Request RW CombinedStatus

The combined status for a specific commit See https://developer.github.com/v3/repos/statuses/#get-the-combined-status-for-a-specific-ref

Data definitions

module GitHub.Data

Request handling

module GitHub.Request