GitHub.Data.Definitions

Documentation

data Error

Errors have been tagged according to their source, so you can more easily dispatch and handle them.

Constructors

HTTPError !HttpException
A HTTP error occurred. The actual caught error is included.

ParseError !Text
An error in the parser itself.

JsonError !Text
The JSON is malformed or unexpected.

UserError !Text
Incorrect input.

data OwnerType

Type of the repository owners.

Constructors

OwnerUser

OwnerOrganization

OwnerBot

data SimpleUser

Constructors

SimpleUser

Fields

simpleUserId :: !(Id User)

simpleUserLogin :: !(Name User)

simpleUserAvatarUrl :: !URL

simpleUserUrl :: !URL

data SimpleOrganization

Constructors

SimpleOrganization

Fields

simpleOrganizationId :: !(Id Organization)

simpleOrganizationLogin :: !(Name Organization)

simpleOrganizationUrl :: !URL

simpleOrganizationAvatarUrl :: !URL

data SimpleOwner

Sometimes we don't know the type of the owner, e.g. in Repo

Constructors

SimpleOwner

Fields

simpleOwnerId :: !(Id Owner)

simpleOwnerLogin :: !(Name Owner)

simpleOwnerUrl :: !URL

simpleOwnerAvatarUrl :: !URL

simpleOwnerType :: !OwnerType

data User

Constructors

User

Fields

userId :: !(Id User)

userLogin :: !(Name User)

userName :: !(Maybe Text)

userType :: !OwnerType

Should always be OwnerUser or OwnerBot

userCreatedAt :: !UTCTime

userPublicGists :: !Int

userAvatarUrl :: !URL

userFollowers :: !Int

userFollowing :: !Int

userHireable :: !(Maybe Bool)

userBlog :: !(Maybe Text)

userBio :: !(Maybe Text)

userPublicRepos :: !Int

userLocation :: !(Maybe Text)

userCompany :: !(Maybe Text)

userEmail :: !(Maybe Text)

userUrl :: !URL

userHtmlUrl :: !URL

data Organization

Constructors

Organization

Fields

organizationId :: !(Id Organization)

organizationLogin :: !(Name Organization)

organizationName :: !(Maybe Text)

organizationType :: !OwnerType

Should always be OwnerOrganization

organizationBlog :: !(Maybe Text)

organizationLocation :: !(Maybe Text)

organizationFollowers :: !Int

organizationCompany :: !(Maybe Text)

organizationAvatarUrl :: !URL

organizationPublicGists :: !Int

organizationHtmlUrl :: !URL

organizationEmail :: !(Maybe Text)

organizationFollowing :: !Int

organizationPublicRepos :: !Int

organizationUrl :: !URL

organizationCreatedAt :: !UTCTime

newtype Owner

In practice you can't have concrete values of Owner.

Constructors

Owner (Either User Organization)

fromOwner :: Owner -> Either User Organization

parseUser :: Object -> Parser User

parseOrganization :: Object -> Parser Organization

data OrgMemberFilter

Filter members returned in the list.

Constructors

OrgMemberFilter2faDisabled
Members without two-factor authentication enabled. Available for organization owners.

OrgMemberFilterAll
All members the authenticated user can see.

data OrgMemberRole

Filter members returned by their role.

Constructors

OrgMemberRoleAll
All members of the organization, regardless of role.

OrgMemberRoleAdmin
Organization owners.

OrgMemberRoleMember
Non-owner organization members.

type QueryString = [(ByteString, Maybe ByteString)]

Request query string

type Count = Int

Count of elements

newtype IssueNumber

Constructors

IssueNumber Int

unIssueNumber :: IssueNumber -> Int

data IssueLabel

Constructors

IssueLabel

Fields

labelColor :: !Text

labelUrl :: !URL

labelName :: !(Name IssueLabel)

labelDesc :: !(Maybe Text)

data NewIssueLabel

Constructors

NewIssueLabel

Fields

newLabelColor :: !Text

newLabelName :: !(Name NewIssueLabel)

newLabelDesc :: !(Maybe Text)

data UpdateIssueLabel

Constructors

UpdateIssueLabel

Fields

updateLabelColor :: !Text

updateLabelName :: !(Name UpdateIssueLabel)

updateLabelDesc :: !(Maybe Text)