GitHub.Data.Repos

Description

This module also exports FromJSON a => FromJSON (HashMap Language a) orphan-ish instance for aeson < 1

Documentation

data Repo

Constructors

Repo

Fields

repoSshUrl :: !(Maybe URL)

repoDescription :: !(Maybe Text)

repoCreatedAt :: !(Maybe UTCTime)

repoHtmlUrl :: !URL

repoSvnUrl :: !(Maybe URL)

repoForks :: !(Maybe Int)

repoHomepage :: !(Maybe Text)

repoFork :: !(Maybe Bool)

repoGitUrl :: !(Maybe URL)

repoPrivate :: !Bool

repoArchived :: !Bool

repoCloneUrl :: !(Maybe URL)

repoSize :: !(Maybe Int)

repoUpdatedAt :: !(Maybe UTCTime)

repoWatchers :: !(Maybe Int)

repoOwner :: !SimpleOwner

repoName :: !(Name Repo)

repoLanguage :: !(Maybe Language)

repoDefaultBranch :: !(Maybe Text)

repoPushedAt :: !(Maybe UTCTime)

this is Nothing for new repositories

repoId :: !(Id Repo)

repoUrl :: !URL

repoOpenIssues :: !(Maybe Int)

repoHasWiki :: !(Maybe Bool)

repoHasIssues :: !(Maybe Bool)

repoHasDownloads :: !(Maybe Bool)

repoParent :: !(Maybe RepoRef)

repoSource :: !(Maybe RepoRef)

repoHooksUrl :: !URL

repoStargazersCount :: !Int

data RepoRef

Constructors

RepoRef

Fields

repoRefOwner :: !SimpleOwner

repoRefRepo :: !(Name Repo)

data NewRepo

Constructors

NewRepo

Fields

newRepoName :: !(Name Repo)

newRepoDescription :: !(Maybe Text)

newRepoHomepage :: !(Maybe Text)

newRepoPrivate :: !(Maybe Bool)

newRepoHasIssues :: !(Maybe Bool)

newRepoHasWiki :: !(Maybe Bool)

newRepoAutoInit :: !(Maybe Bool)

newRepo :: Name Repo -> NewRepo

data EditRepo

Constructors

EditRepo

Fields

editName :: !(Maybe (Name Repo))

editDescription :: !(Maybe Text)

editHomepage :: !(Maybe Text)

editPublic :: !(Maybe Bool)

editHasIssues :: !(Maybe Bool)

editHasWiki :: !(Maybe Bool)

editHasDownloads :: !(Maybe Bool)

data RepoPublicity

Filter the list of the user's repos using any of these constructors.

Constructors

RepoPublicityAll
All repos accessible to the user.

RepoPublicityOwner
Only repos owned by the user.

RepoPublicityPublic
Only public repos.

RepoPublicityPrivate
Only private repos.

RepoPublicityMember
Only repos to which the user is a member but not an owner.

type Languages = HashMap Language Int

The value is the number of bytes of code written in that language.

newtype Language

A programming language.

Constructors

Language Text

getLanguage :: Language -> Text

data Contributor

Constructors

KnownContributor !Int !URL !(Name User) !URL !(Id User) !Text
An existing Github user, with their number of contributions, avatar URL, login, URL, ID, and Gravatar ID.

AnonymousContributor !Int !Text
An unknown Github user with their number of contributions and recorded name.

contributorToSimpleUser :: Contributor -> Maybe SimpleUser

data ArchiveFormat

Constructors

ArchiveFormatTarball
".tar.gz" format

ArchiveFormatZipball
".zip" format