GitHub.Data.Request

Description

Request

data Request (k :: RW) a where

Github request data type.

k describes whether authentication is required. It's required for non-GET requests.

a is the result type

Note: Request is not Functor on purpose.

Constructors

SimpleQuery :: FromJSON a => SimpleRequest k a -> Request k a

StatusQuery :: StatusMap a -> SimpleRequest k () -> Request k a

HeaderQuery :: FromJSON a => RequestHeaders -> SimpleRequest k a -> Request k a

RedirectQuery :: SimpleRequest k () -> Request k URI

data SimpleRequest (k :: RW) a where

Constructors

Query :: Paths -> QueryString -> SimpleRequest k a

PagedQuery :: Paths -> QueryString -> FetchCount -> SimpleRequest k (Vector a)

Command :: CommandMethod a -> Paths -> ByteString -> SimpleRequest RW a

Smart constructors

query :: FromJSON a => Paths -> QueryString -> Request k a

pagedQuery :: FromJSON a => Paths -> QueryString -> FetchCount -> Request k (Vector a)

command :: FromJSON a => CommandMethod a -> Paths -> ByteString -> Request RW a

Auxiliary types

data RW

Type used as with DataKinds to tag whether requests need authentication or aren't read-only.

Constructors

RO
Read-only, doesn't necessarily requires authentication

RA
Read autenticated

RW
Read-write, requires authentication

type StatusMap a = [(Int, a)]

statusOnlyOk :: StatusMap Bool

data CommandMethod a where

Http method of requests with body.

Constructors

Post :: CommandMethod a

Patch :: CommandMethod a

Put :: CommandMethod a

Put' :: CommandMethod ()

Delete :: CommandMethod ()

toMethod :: CommandMethod a -> Method

data FetchCount

PagedQuery returns just some results, using this data we can specify how many pages we want to fetch.

Constructors

FetchAtLeast !Word

FetchAll

type Paths = [Text]

class IsPathPart a where

Minimal complete definition

toPathPart

Methods

toPathPart :: a -> Text

type QueryString = [(ByteString, Maybe ByteString)]

Request query string

type Count = Int

Count of elements