GitHub.Endpoints.Issues.Labels

Description

The API for dealing with labels on Github issues as described on http://developer.github.com/v3/issues/labels/.

Documentation

labelsOnRepo :: Name Owner -> Name Repo -> IO (Either Error (Vector IssueLabel))

All the labels available to use on any issue in the repo.

labelsOnRepo "thoughtbot" "paperclip"

labelsOnRepo' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector IssueLabel))

All the labels available to use on any issue in the repo using authentication.

labelsOnRepo' (Just (User (user password))) "thoughtbot" "paperclip"

labelsOnRepoR :: Name Owner -> Name Repo -> FetchCount -> Request k (Vector IssueLabel)

List all labels for this repository. See https://developer.github.com/v3/issues/labels/#list-all-labels-for-this-repository

label :: Name Owner -> Name Repo -> Name IssueLabel -> IO (Either Error IssueLabel)

A label by name.

label "thoughtbot" "paperclip" "bug"

label' :: Maybe Auth -> Name Owner -> Name Repo -> Name IssueLabel -> IO (Either Error IssueLabel)

A label by name using authentication.

label' (Just (User (user password))) "thoughtbot" "paperclip" "bug"

labelR :: Name Owner -> Name Repo -> Name IssueLabel -> Request k IssueLabel

Query a single label. See https://developer.github.com/v3/issues/labels/#get-a-single-label

createLabel :: Auth -> Name Owner -> Name Repo -> Name IssueLabel -> String -> IO (Either Error IssueLabel)

Create a label

createLabel (User (user password)) "thoughtbot" "paperclip" "bug" "f29513"

createLabelR :: Name Owner -> Name Repo -> Name IssueLabel -> String -> Request RW IssueLabel

Create a label. See https://developer.github.com/v3/issues/labels/#create-a-label

updateLabel

Arguments

:: Auth

-> Name Owner

-> Name Repo

-> Name IssueLabel
old label name

-> Name IssueLabel
new label name

-> String
new color

-> IO (Either Error IssueLabel)

Update a label

updateLabel (User (user password)) "thoughtbot" "paperclip" "bug" "new-bug" "ff1111"

updateLabelR

Arguments

:: Name Owner

-> Name Repo

-> Name IssueLabel
old label name

-> Name IssueLabel
new label name

-> String
new color

-> Request RW IssueLabel

Update a label. See https://developer.github.com/v3/issues/labels/#update-a-label

deleteLabel :: Auth -> Name Owner -> Name Repo -> Name IssueLabel -> IO (Either Error ())

Delete a label

deleteLabel (User (user password)) "thoughtbot" "paperclip" "bug"

deleteLabelR :: Name Owner -> Name Repo -> Name IssueLabel -> Request RW ()

Delete a label. See https://developer.github.com/v3/issues/labels/#delete-a-label

labelsOnIssue :: Name Owner -> Name Repo -> Id Issue -> IO (Either Error (Vector IssueLabel))

The labels on an issue in a repo.

labelsOnIssue "thoughtbot" "paperclip" 585

labelsOnIssue' :: Maybe Auth -> Name Owner -> Name Repo -> Id Issue -> IO (Either Error (Vector IssueLabel))

The labels on an issue in a repo using authentication.

labelsOnIssue' (Just (User (user password))) "thoughtbot" "paperclip" (Id 585)

labelsOnIssueR :: Name Owner -> Name Repo -> Id Issue -> FetchCount -> Request k (Vector IssueLabel)

List labels on an issue. See https://developer.github.com/v3/issues/labels/#list-labels-on-an-issue

addLabelsToIssue :: Foldable f => Auth -> Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> IO (Either Error (Vector IssueLabel))

Add labels to an issue.

addLabelsToIssue (User (user password)) "thoughtbot" "paperclip" (Id 585) ["Label1" "Label2"]

addLabelsToIssueR :: Foldable f => Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> Request RW (Vector IssueLabel)

Add lables to an issue. See https://developer.github.com/v3/issues/labels/#add-labels-to-an-issue

removeLabelFromIssue :: Auth -> Name Owner -> Name Repo -> Id Issue -> Name IssueLabel -> IO (Either Error ())

Remove a label from an issue.

removeLabelFromIssue (User (user password)) "thoughtbot" "paperclip" (Id 585) "bug"

removeLabelFromIssueR :: Name Owner -> Name Repo -> Id Issue -> Name IssueLabel -> Request RW ()

Remove a label from an issue. See https://developer.github.com/v3/issues/labels/#remove-a-label-from-an-issue

replaceAllLabelsForIssue :: Foldable f => Auth -> Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> IO (Either Error (Vector IssueLabel))

Replace all labels on an issue. Sending an empty list will remove all labels from the issue.

replaceAllLabelsForIssue (User (user password)) "thoughtbot" "paperclip" (Id 585) ["Label1" "Label2"]

replaceAllLabelsForIssueR :: Foldable f => Name Owner -> Name Repo -> Id Issue -> f (Name IssueLabel) -> Request RW (Vector IssueLabel)

Replace all labels on an issue. See https://developer.github.com/v3/issues/labels/#replace-all-labels-for-an-issue

Sending an empty list will remove all labels from the issue.

removeAllLabelsFromIssue :: Auth -> Name Owner -> Name Repo -> Id Issue -> IO (Either Error ())

Remove all labels from an issue.

removeAllLabelsFromIssue (User (user password)) "thoughtbot" "paperclip" (Id 585)

removeAllLabelsFromIssueR :: Name Owner -> Name Repo -> Id Issue -> Request RW ()

Remove all labels from an issue. See https://developer.github.com/v3/issues/labels/#remove-all-labels-from-an-issue

labelsOnMilestone :: Name Owner -> Name Repo -> Id Milestone -> IO (Either Error (Vector IssueLabel))

All the labels on a repo's milestone given the milestone ID.

labelsOnMilestone "thoughtbot" "paperclip" (Id 2)

labelsOnMilestone' :: Maybe Auth -> Name Owner -> Name Repo -> Id Milestone -> IO (Either Error (Vector IssueLabel))

All the labels on a repo's milestone given the milestone ID using authentication.

labelsOnMilestone' (Just (User (user password))) "thoughtbot" "paperclip" (Id 2)

labelsOnMilestoneR :: Name Owner -> Name Repo -> Id Milestone -> FetchCount -> Request k (Vector IssueLabel)

Query labels for every issue in a milestone. See https://developer.github.com/v3/issues/labels/#get-labels-for-every-issue-in-a-milestone

module GitHub.Data