GitHub.Endpoints.Organizations.Members

Description

The organization members API as described on http://developer.github.com/v3/orgs/members/.

Documentation

membersOf :: Name Organization -> IO (Either Error (Vector SimpleUser))

All the users who are members of the specified organization, | without authentication.

membersOf "thoughtbot"

membersOf' :: Maybe Auth -> Name Organization -> IO (Either Error (Vector SimpleUser))

All the users who are members of the specified organization, | with or without authentication.

membersOf' (Just $ OAuth "token") "thoughtbot"

membersOfR :: Name Organization -> FetchCount -> Request k (Vector SimpleUser)

All the users who are members of the specified organization.

See https://developer.github.com/v3/orgs/members/#members-list

membersOfWithR :: Name Organization -> OrgMemberFilter -> OrgMemberRole -> FetchCount -> Request k (Vector SimpleUser)

membersOfR with filters.

See https://developer.github.com/v3/orgs/members/#members-list

isMemberOf :: Name User -> Name Organization -> IO (Either Error Bool)

Check if a user is a member of an organization, | without authentication.

isMemberOf "phadej" "haskell-infra"

isMemberOf' :: Maybe Auth -> Name User -> Name Organization -> IO (Either Error Bool)

Check if a user is a member of an organization, | with or without authentication.

isMemberOf' (Just $ OAuth "token") "phadej" "haskell-infra"

isMemberOfR :: Name User -> Name Organization -> Request k Bool

Check if a user is a member of an organization.

See https://developer.github.com/v3/orgs/members/#check-membership

orgInvitationsR :: Name Organization -> FetchCount -> Request RA (Vector Invitation)

List pending organization invitations

See https://developer.github.com/v3/orgs/members/#list-pending-organization-invitations

module GitHub.Data