GitHub.Endpoints.PullRequests

Description

The pull requests API as documented at http://developer.github.com/v3/pulls/.

Documentation

pullRequestsFor :: Name Owner -> Name Repo -> IO (Either Error (Vector SimplePullRequest))

All open pull requests for the repo, by owner and repo name.

pullRequestsFor "rails" "rails"

pullRequestsFor' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector SimplePullRequest))

All open pull requests for the repo, by owner and repo name.

pullRequestsFor "rails" "rails"

pullRequestsForR :: Name Owner -> Name Repo -> PullRequestMod -> FetchCount -> Request k (Vector SimplePullRequest)

List pull requests. See https://developer.github.com/v3/pulls/#list-pull-requests

pullRequest' :: Maybe Auth -> Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error PullRequest)

A detailed pull request, which has much more information. This takes the repo owner and name along with the number assigned to the pull request. With authentification.

pullRequest' (Just ("github-username", "github-password")) "thoughtbot" "paperclip" 562

pullRequest :: Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error PullRequest)

A detailed pull request, which has much more information. This takes the repo owner and name along with the number assigned to the pull request.

pullRequest "thoughtbot" "paperclip" 562

pullRequestR :: Name Owner -> Name Repo -> Id PullRequest -> Request k PullRequest

Query a single pull request. See https://developer.github.com/v3/pulls/#get-a-single-pull-request

createPullRequest :: Auth -> Name Owner -> Name Repo -> CreatePullRequest -> IO (Either Error PullRequest)

createPullRequestR :: Name Owner -> Name Repo -> CreatePullRequest -> Request RW PullRequest

Create a pull request. See https://developer.github.com/v3/pulls/#create-a-pull-request

updatePullRequest :: Auth -> Name Owner -> Name Repo -> Id PullRequest -> EditPullRequest -> IO (Either Error PullRequest)

Update a pull request

updatePullRequestR :: Name Owner -> Name Repo -> Id PullRequest -> EditPullRequest -> Request RW PullRequest

Update a pull request. See https://developer.github.com/v3/pulls/#update-a-pull-request

pullRequestCommits' :: Maybe Auth -> Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error (Vector Commit))

All the commits on a pull request, given the repo owner, repo name, and the number of the pull request. With authentification.

pullRequestCommits' (Just ("github-username", "github-password")) "thoughtbot" "paperclip" 688

pullRequestCommitsIO :: Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error (Vector Commit))

All the commits on a pull request, given the repo owner, repo name, and the number of the pull request.

pullRequestCommits "thoughtbot" "paperclip" 688

pullRequestCommitsR :: Name Owner -> Name Repo -> Id PullRequest -> FetchCount -> Request k (Vector Commit)

List commits on a pull request. See https://developer.github.com/v3/pulls/#list-commits-on-a-pull-request

pullRequestFiles' :: Maybe Auth -> Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error (Vector File))

The individual files that a pull request patches. Takes the repo owner and name, plus the number assigned to the pull request. With authentification.

pullRequestFiles' (Just ("github-username", "github-password")) "thoughtbot" "paperclip" 688

pullRequestFiles :: Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error (Vector File))

The individual files that a pull request patches. Takes the repo owner and name, plus the number assigned to the pull request.

pullRequestFiles "thoughtbot" "paperclip" 688

pullRequestFilesR :: Name Owner -> Name Repo -> Id PullRequest -> FetchCount -> Request k (Vector File)

List pull requests files. See https://developer.github.com/v3/pulls/#list-pull-requests-files

isPullRequestMerged :: Auth -> Name Owner -> Name Repo -> Id PullRequest -> IO (Either Error Bool)

Check if pull request has been merged.

isPullRequestMergedR :: Name Owner -> Name Repo -> Id PullRequest -> Request k Bool

Query if a pull request has been merged. See https://developer.github.com/v3/pulls/#get-if-a-pull-request-has-been-merged

mergePullRequest :: Auth -> Name Owner -> Name Repo -> Id PullRequest -> Maybe Text -> IO (Either Error MergeResult)

Merge a pull request.

mergePullRequestR :: Name Owner -> Name Repo -> Id PullRequest -> Maybe Text -> Request RW MergeResult

Merge a pull request (Merge Button). https://developer.github.com/v3/pulls/#merge-a-pull-request-merge-button

module GitHub.Data