GitHub.Request

Description

This module provides data types and helper methods, which makes possible to build alternative API request intepreters in addition to provided IO functions.

Simple example using operational package. See samples/Operational/Operational.hs

type GithubMonad a = Program (GH.Request 'False) a

-- | Intepret GithubMonad value into IO
runMonad :: Manager -> GH.Auth -> GithubMonad a -> ExceptT GH.Error IO a
runMonad mgr auth m = case view m of
   Return a   -> return a
   req :>>= k -> do
       b <- ExceptT $ GH.executeRequestWithMgr mgr auth req
       runMonad mgr auth (k b)

-- | Lift request into Monad
githubRequest :: GH.Request 'False a -> GithubMonad a
githubRequest = singleton

Types

data Request (k :: RW) a where

Github request data type.

k describes whether authentication is required. It's required for non-GET requests.

a is the result type

Note: Request is not Functor on purpose.

Constructors

SimpleQuery :: FromJSON a => SimpleRequest k a -> Request k a

StatusQuery :: StatusMap a -> SimpleRequest k () -> Request k a

HeaderQuery :: FromJSON a => RequestHeaders -> SimpleRequest k a -> Request k a

RedirectQuery :: SimpleRequest k () -> Request k URI

data CommandMethod a where

Http method of requests with body.

Constructors

Post :: CommandMethod a

Patch :: CommandMethod a

Put :: CommandMethod a

Put' :: CommandMethod ()

Delete :: CommandMethod ()

toMethod :: CommandMethod a -> Method

type Paths = [Text]

type QueryString = [(ByteString, Maybe ByteString)]

Request query string

Request execution in IO

executeRequest :: Auth -> Request k a -> IO (Either Error a)

Execute Request in IO

executeRequestWithMgr :: Manager -> Auth -> Request k a -> IO (Either Error a)

Like executeRequest but with provided Manager.

executeRequest' :: Request RO a -> IO (Either Error a)

Like executeRequest but without authentication.

executeRequestWithMgr' :: Manager -> Request RO a -> IO (Either Error a)

Like executeRequestWithMgr but without authentication.

executeRequestMaybe :: Maybe Auth -> Request RO a -> IO (Either Error a)

Helper for picking between executeRequest and executeRequest'.

The use is discouraged.

unsafeDropAuthRequirements :: Request k' a -> Request k a

Partial function to drop authentication need.

Helpers

makeHttpRequest :: MonadThrow m => Maybe Auth -> Request k a -> m Request

Create http-client Request.

for PagedQuery, the initial request is created.

for Status, the Request for underlying Request is created, status checking is modifying accordingly.

parseResponse :: Maybe Auth -> Request k a -> Maybe Request

makeHttpSimpleRequest :: MonadThrow m => Maybe Auth -> SimpleRequest k a -> m Request

parseResponse :: (FromJSON a, MonadError Error m) => Response ByteString -> m a

Parse API response.

parseResponse :: FromJSON a => Response ByteString -> Either Error a

parseStatus :: MonadError Error m => StatusMap a -> Status -> m a

Helper for handling of RequestStatus.

parseStatus :: StatusMap a -> Status -> Either Error a

getNextUrl :: Response a -> Maybe URI

Query Link header with rel=next from the request headers.

performPagedRequest

Arguments

:: (FromJSON a, Semigroup a, MonadCatch m, MonadError Error m)

=> (Request -> m (Response ByteString))
httpLbs analogue

-> (a -> Bool)
predicate to continue iteration

-> Request
initial request

-> m a

Helper for making paginated requests. Responses, a are combined monoidally.

performPagedRequest :: (FromJSON a, Semigroup a)
                    => (Request -> ExceptT Error IO (Response ByteString))
                    -> (a -> Bool)
                    -> Request
                    -> ExceptT Error IO a